% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ckanr_settings.R
\name{api_key}
\alias{api_key}
\title{Enable CKAN API authentication with an API key}
\usage{
api_key(key = NULL)
}
\description{
Some CKAN API functions require authentication through the HTTP request header
\code{X-CKAN-API-Key}. \code{ckanR} stores the API key in an environment
variable \code{CKANR_DEFAULT_KEY}, set by \code{setup_ckanr},
to avoid namespace collisions.
\code{api_key} facilitates authentication by adding the configured API key
as the correctly named request header. It serves as default value in all
\code{ckanr} functions for convenience.
}
\keyword{internal}

