% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-ckan_file.R
\name{as.ckan_file}
\alias{as.ckan_file}
\alias{is.ckan_file}
\title{ckan_file class helpers}
\usage{
as.ckan_file(x, ...)

is.ckan_file(x)
}
\arguments{
\item{x}{One of character, list, or ckan_file class object}

\item{...}{Extra arguments. If \code{x} is character, the function passes them
on to \code{\link[=file_show]{file_show()}}}
}
\description{
ckan_file class helpers
}
\examples{
\dontrun{
ckanr_setup(
  url = "https://demo.ckan.org/",
  key = getOption("ckan_demo_key")
)

# create item class from only an item ID
as.ckan_file("file-id-here")

# gives back itself
(x <- as.ckan_file(list(id = "file-id-here", name = "file.txt")))
as.ckan_file(x)
}
}
