% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-ckan_group.R
\name{as.ckan_group}
\alias{as.ckan_group}
\alias{is.ckan_group}
\title{ckan_group class helpers}
\usage{
as.ckan_group(x, ...)

is.ckan_group(x)
}
\arguments{
\item{x}{One of character, list, or ckan_group class object}

\item{...}{Extra arguments. If \code{x} is character, the function passes them
on to \code{\link[=group_show]{group_show()}}}
}
\description{
ckan_group class helpers
}
\examples{
\dontrun{
ckanr_setup(url = "https://demo.ckan.org/", key = getOption("ckan_demo_key"))

(grps <- group_list())
grps[[3]]

# create item class from only an item ID
as.ckan_group(grps[[3]]$id)

# gives back itself
(x <- as.ckan_group(grps[[3]]$id))
as.ckan_group(x)
}
}
