% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-ckan_package.R
\name{as.ckan_package}
\alias{as.ckan_package}
\alias{is.ckan_package}
\title{ckan_package class helpers}
\usage{
as.ckan_package(x, ...)

is.ckan_package(x)
}
\arguments{
\item{x}{One of character, list, or ckan_package class object}

\item{...}{Extra arguments. If \code{x} is character, the function passes them
on to \code{\link[=package_show]{package_show()}}. If GET is not supported, use the \code{http_method}
parameter to set a different HTTP verb}
}
\description{
ckan_package class helpers
}
\examples{
\dontrun{
ckanr_setup(
  url = "https://demo.ckan.org/",
  key = getOption("ckan_demo_key")
)

(pkgs <- package_search())
pkgs$results
pkgs$results[[3]]

# create item class from only an item ID
as.ckan_package(pkgs$results[[3]]$id)

# gives back itself
(x <- as.ckan_package(pkgs$results[[3]]$id))
as.ckan_package(x)
}
}
