% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-ckan_resource.R
\name{as.ckan_resource}
\alias{as.ckan_resource}
\alias{is.ckan_resource}
\title{ckan_resource class helpers}
\usage{
as.ckan_resource(x, ...)

is.ckan_resource(x)
}
\arguments{
\item{x}{One of character, list, or ckan_package class object}

\item{...}{Extra arguments. If \code{x} is character, the function passes them
on to \code{\link[=resource_show]{resource_show()}}}
}
\description{
ckan_resource class helpers
}
\examples{
\dontrun{
ckanr_setup(
  url = "https://demo.ckan.org/",
  key = getOption("ckan_demo_key")
)

(resrcs <- resource_search(q = "name:data"))
resrcs$results
resrcs$results[[3]]

# create item class from only an item ID
as.ckan_resource(resrcs$results[[3]]$id)

# gives back itself
(x <- as.ckan_resource(resrcs$results[[3]]$id))
as.ckan_resource(x)
}
}
