% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-ckan_resource_view.R
\name{as.ckan_resource_view}
\alias{as.ckan_resource_view}
\alias{is.ckan_resource_view}
\title{ckan_resource_view class helpers}
\usage{
as.ckan_resource_view(x, ...)

is.ckan_resource_view(x)
}
\arguments{
\item{x}{One of character, list, or \code{ckan_resource_view} object}

\item{...}{Extra arguments. If \code{x} is an identifier, the function passes
them to \code{\link[=resource_view_show]{resource_view_show()}}.}
}
\description{
ckan_resource_view class helpers
}
\examples{
\dontrun{
ckanr_setup(
  url = "https://demo.ckan.org/",
  key = Sys.getenv("CKAN_DEMO_KEY")
)

res <- package_show("sample-dataset")
views <- resource_view_list(res$resources[[1]]$id)

# Coerce from ID
as.ckan_resource_view(views[[1]]$id)

# Pass through existing objects
as.ckan_resource_view(views[[1]])
}
}
