% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-ckan_tag.R
\name{as.ckan_tag}
\alias{as.ckan_tag}
\alias{is.ckan_tag}
\title{ckan_tag class helpers}
\usage{
as.ckan_tag(x, ...)

is.ckan_tag(x)
}
\arguments{
\item{x}{One of character, list, or ckan_tag class object}

\item{...}{Extra arguments. If \code{x} is character, the function passes them
on to \code{\link[=tag_show]{tag_show()}}}
}
\description{
ckan_tag class helpers
}
\examples{
\dontrun{
ckanr_setup(
  url = "https://demo.ckan.org/",
  key = getOption("ckan_demo_key")
)

(tags <- tag_search(query = "ta"))
tags[[3]]

# create item class from only an item ID
as.ckan_tag(tags[[3]]$id)

# gives back itself
(x <- as.ckan_tag(tags[[3]]$id))
as.ckan_tag(x)
}
}
