% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-ckan_user.R
\name{as.ckan_user}
\alias{as.ckan_user}
\alias{is.ckan_user}
\title{ckan_user class helpers}
\usage{
as.ckan_user(x, ...)

is.ckan_user(x)
}
\arguments{
\item{x}{One of character, list, or ckan_user class object}

\item{...}{Extra arguments. If \code{x} is character, the function passes them
on to \code{\link[=user_show]{user_show()}}}
}
\description{
ckan_user class helpers
}
\examples{
\dontrun{
ckanr_setup(
  url = "https://demo.ckan.org/",
  key = getOption("ckan_demo_key")
)

(usrs <- user_list())
usrs[1:3]
usrs[[3]]

# create item class from only an item ID
as.ckan_user(usrs[[3]]$id)

# gives back itself
(x <- as.ckan_user(usrs[[3]]$id))
as.ckan_user(x)
}
}
