% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-ckan_vocabulary.R
\name{as.ckan_vocabulary}
\alias{as.ckan_vocabulary}
\alias{is.ckan_vocabulary}
\title{ckan_vocabulary class helpers}
\usage{
as.ckan_vocabulary(x, ...)

is.ckan_vocabulary(x)
}
\arguments{
\item{x}{One of character, list, or ckan_vocabulary object}

\item{...}{Extra arguments. If you retrieve by identifier, the function
passes them to \code{vocabulary_show()}.}
}
\description{
ckan_vocabulary class helpers
}
\examples{
\dontrun{
ckanr_setup(url = "https://demo.ckan.org/", key = getOption("ckan_demo_key"))

vocab <- vocabulary_create(name = sprintf("demo_vocab_\%s", sample.int(1e4, 1)))

# create class from an id
as.ckan_vocabulary(vocab$id)

# passing through existing objects is a no-op
as.ckan_vocabulary(vocab)
}
}
