% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-action.R
\name{ckan_action}
\alias{ckan_action}
\title{Generic action function}
\usage{
ckan_action(
  action,
  query = NULL,
  body = NULL,
  headers = list(),
  verb = "POST",
  url = get_default_url(),
  key = get_default_key(),
  ...
)
}
\arguments{
\item{action}{A valid CKAN API action name (for example, "package_list", "package_show"). See the CKAN API documentation for a full list of actions: https://docs.ckan.org/en/latest/api/index.html}

\item{query}{a named list of URL query parameters}

\item{body}{Data for the body of a request. See
https://docs.ropensci.org/crul/reference/verb-POST.html for options}

\item{headers}{a named list of request headers}

\item{verb}{HTTP request verb, for example, GET, POST}

\item{url}{Base url to use. Default: https://demo.ckan.org/ See
also \code{\link[=ckanr_setup]{ckanr_setup()}} and \code{\link[=get_default_url]{get_default_url()}}}

\item{key}{A privileged CKAN API key. Default: your key set with
\code{\link{ckanr_setup}}}

\item{...}{Curl args. The function sends them to the relevant \link[crul:HttpClient]{crul::HttpClient} method for the \code{verb} parameter (optional)}
}
\value{
A text string. The function returns text because the data type is unknown ahead of time. You parse the text as needed.
}
\description{
You must set the CKAN action and the HTTP verb.
You also set query parameters, body data, and headers.
}
\examples{
\dontrun{
ckanr_setup(
  url = "https://demo.ckan.org/",
  key = getOption("ckan_demo_key")
)

ckan_action("package_list")
ckan_action("package_list", verb = "GET")
ckan_action("package_list", url = "https://data.nhm.ac.uk")
}
}
