% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckan_classes.R
\name{ckan_classes}
\alias{ckan_classes}
\title{ckanr S3 classes}
\description{
ckanr S3 classes
}
\section{The classes}{

\itemize{
\item ckan_package - CKAN package
\item ckan_resource - CKAN resource
\item ckan_resource_view - CKAN resource view
\item ckan_related - CKAN related item
}
}

\section{Coercion}{

The \verb{as.ckan_*()} functions for each CKAN object type coerce an object
to an S3 class of that type. For example, you can coerce a package ID as a
character string into an \code{ckan_package} object by calling
\verb{as.ckan_package(<id>}.
}

\section{Testing for classes}{

To test the class of an object, use an \verb{is._ckan_*()} function for the
classes listed above. The function returns \code{TRUE} or \code{FALSE}.
}

\section{Manipulation}{

These are S3 classes. Each class is an R list with an attached class.
The class tells functions what to do with the object. Functions accept
flexible inputs and return flexible outputs. You can edit one of these
classes yourself by changing values in the list.
}

