% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckan_info.R
\name{ckan_info}
\alias{ckan_info}
\alias{ckan_version}
\title{Get information on a CKAN server}
\usage{
ckan_info(url = get_default_url(), ...)

ckan_version(url = get_default_url(), ...)
}
\arguments{
\item{url}{Base URL to use. Default: \url{https://demo.ckan.org/}. See
also \code{\link[=ckanr_setup]{ckanr_setup()}} and \code{\link[=get_default_url]{get_default_url()}}. (required)}

\item{...}{Extra curl arguments. The function passes them on
to \link[crul:verb-GET]{crul::verb-GET} (optional)}
}
\value{
For \code{ckan_info}, the function returns a list with many slots with
various info. For \code{ckan_version}, the function returns a list of length two,
with the actual version as character. The second item converts the version
to numeric (any dots or letters removed)
}
\description{
Get information on a CKAN server
}
\examples{
\dontrun{
ckan_info()
ckan_info(servers()[5])

ckan_version(servers()[5])
}
}
