% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zzz.R
\name{ckan_POST}
\alias{ckan_POST}
\title{Send an httr::POST request and handle the response}
\usage{
ckan_POST(url, method, body = NULL, key = NULL, ...)
}
\arguments{
\item{url}{A CKAN base URL}

\item{method}{The GET method as part of the CKAN API URL}

\item{body}{The request body (a dictionary as named R list) (optional)}

\item{key}{A CKAN API key (optional)}
}
\value{
The content of the response as text
}
\description{
Send an httr::POST request and handle the response
}
\details{
This method is a thin wrapper around httr::POST, which is a wrapper
around RCurl::POST, which is a wrapper around curl.

If the POST request requires authentication, a privileged CKAN API key has
to be supplied. The method does not default to \code{ckanr_settings}.
As the method is an internal method only, other functions of \code{ckanr}
may default to \code{ckanr_settings} for convenience of use.
}
\keyword{internal}

