% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_ops.R
\name{config_options}
\alias{config_options}
\alias{config_option_list}
\alias{config_option_show}
\alias{config_option_update}
\title{Runtime configuration helpers}
\usage{
config_option_list(
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)

config_option_show(
  option_key,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)

config_option_update(
  options,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\arguments{
\item{url}{Base URL to use. Default: https://demo.ckan.org/. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key. Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
the table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible. The result
can vary from a vector, list or data.frame. (required)}

\item{...}{Extra curl arguments. The function passes them to \code{\link[crul]{verb-POST}} (optional)}

\item{option_key}{(character) Configuration option identifier.}

\item{options}{(list) Named list of configuration options to update.}
}
\description{
Inspect and update CKAN's runtime-editable configuration keys.
}
\examples{
\dontrun{
ckanr_setup(url = "https://demo.ckan.org/", key = "my-ckan-key")
config_option_list()
config_option_update(list("ckan.site_title" = "My Portal"))
}
}
