% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_count.R
\name{dashboard_count}
\alias{dashboard_count}
\title{Number of new activities of an authorized user}
\usage{
dashboard_count(
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\arguments{
\item{url}{Base URL to use. Default: https://demo.ckan.org/. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key. Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
the table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible. The result
can vary from a vector, list or data.frame. (required)}

\item{...}{Extra curl arguments. The function passes them to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
Number of new activities of an authorized user
}
\details{
CKAN does not count activities from the user herself, but they
appear in the dashboard. Users do not want notices about things they did
themselves.
}
\examples{
\dontrun{
# Setup
ckanr_setup(url = "https://demo.ckan.org/", key = getOption("ckan_demo_key"))

# count
dashboard_new_activities_count()
}
}
