% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_dataset_extras.R
\name{dataset_purge}
\alias{dataset_purge}
\title{Permanently purge a dataset}
\usage{
dataset_purge(id, url = get_default_url(), key = get_default_key(), ...)
}
\arguments{
\item{id}{(character or \code{ckan_package}) Dataset identifier.}

\item{url}{Base URL to use. Default: https://demo.ckan.org/. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key. Default: your key set with
\code{\link{ckanr_setup}}}

\item{...}{Extra curl arguments. The function passes them to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
Permanently purge a dataset
}
\examples{
\dontrun{
dataset_purge("dataset-id")
}
}
