% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{dbplyr_edition.myConnectionClass}
\alias{dbplyr_edition.myConnectionClass}
\title{Use dbplyr 2.0.0 generics
See https://dbplyr.tidyverse.org/articles/backend-2.html#nd-edition
New in 0.7.0.9001}
\usage{
\method{dbplyr_edition}{myConnectionClass}(con)
}
\description{
Use dbplyr 2.0.0 generics
See https://dbplyr.tidyverse.org/articles/backend-2.html#nd-edition
New in 0.7.0.9001
}
