% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{dbplyr_edition.src_ckan}
\alias{dbplyr_edition.src_ckan}
\title{Use dbplyr 2.0.0 generics
See https://dbplyr.tidyverse.org/articles/backend-2.html#nd-edition
Introduced in 0.8.0}
\usage{
\method{dbplyr_edition}{src_ckan}(con)
}
\description{
Use dbplyr 2.0.0 generics
See https://dbplyr.tidyverse.org/articles/backend-2.html#nd-edition
Introduced in 0.8.0
}
