% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_ops.R
\name{diagnostics}
\alias{diagnostics}
\alias{status_show}
\alias{help_show}
\title{Diagnostics helpers}
\usage{
status_show(url = get_default_url(), key = get_default_key(), as = "list", ...)

help_show(
  name,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\arguments{
\item{url}{Base URL to use. Default: https://demo.ckan.org/. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key. Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
the table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible. The result
can vary from a vector, list or data.frame. (required)}

\item{...}{Extra curl arguments. The function passes them to \code{\link[crul]{verb-POST}} (optional)}

\item{name}{(character) CKAN action name to describe.}
}
\description{
Lightweight wrappers for CKAN's \code{status_show} and \code{help_show} actions.
}
\examples{
\dontrun{
ckanr_setup(url = "https://demo.ckan.org/")
status_show()
help_show("package_search")
}
}
