% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_search_sql.R
\name{ds_search_sql}
\alias{ds_search_sql}
\title{Datastore - search or get a dataset from CKAN datastore}
\usage{
ds_search_sql(
  sql,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\arguments{
\item{sql}{(character) A single SQL select statement. (required)}

\item{url}{Base url to use. Default: https://demo.ckan.org/ See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key, Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible, so the result
can vary from a vector, list or data.frame. (required)}

\item{...}{Curl args passed on to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
Datastore - search or get a dataset from CKAN datastore
}
\examples{
\dontrun{
ckanr_setup(url = "https://data.gov.au/")
sql <- 'SELECT * from "eef6a84b-ad44-446f-9cf9-fb5d135e3123" LIMIT 2'
ds_search_sql(sql, as = "table")
sql2 <- 'SELECT "Dog","TRI" from "eef6a84b-ad44-446f-9cf9-fb5d135e3123" LIMIT 2'
ds_search_sql(sql2, as = "table")
}
}
