% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_search_sql_post.R
\name{ds_search_sql_post}
\alias{ds_search_sql_post}
\title{Datastore - search or get a dataset from CKRAN datastore}
\usage{
ds_search_sql_post(sql, url = get_default_url(), as = "list",
  config = httr::add_headers(Authorization = get_default_key(),
  `Content-Type` = "application/json"), ...)
}
\arguments{
\item{sql}{(character) A single SQL select statement. (required)}

\item{url}{Base url to use. Default: \url{http://data.techno-science.ca}. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{as}{(character) One of list (default), table, or json. Parsing with table option
uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)}, which attempts to parse
data to data.frame's when possible, so the result can vary from a vector, list or
data.frame. (required)}

\item{...}{Curl args passed on to \code{\link[httr]{POST}} (optional)}
}
\description{
Datastore - search or get a dataset from CKRAN datastore
}
\examples{
\dontrun{
url <- 'https://demo.ckan.org/'
sql <- 'SELECT * from "f4129802-22aa-4437-b9f9-8a8f3b7b2a53" LIMIT 2'
ds_search_sql_post(sql, url = url, as = "table")
sql2 <- 'SELECT "Species","Genus","Family" from "f4129802-22aa-4437-b9f9-8a8f3b7b2a53" LIMIT 2'
ds_search_sql_post(sql2, url = url, as = "table")
}
}
