% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/followers.R
\name{follow_objects}
\alias{follow_objects}
\alias{dataset_follow}
\alias{dataset_unfollow}
\alias{dataset_am_following}
\alias{group_follow}
\alias{group_unfollow}
\alias{group_am_following}
\title{Follow and unfollow CKAN datasets or groups}
\usage{
dataset_follow(
  id,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)

dataset_unfollow(
  id,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)

dataset_am_following(id, url = get_default_url(), key = get_default_key(), ...)

group_follow(
  id,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)

group_unfollow(
  id,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)

group_am_following(id, url = get_default_url(), key = get_default_key(), ...)
}
\arguments{
\item{id}{(character or \verb{ckan_*}) Identifier for the object to inspect.
For datasets, pass an id/slug or \code{ckan_package}. For groups, pass the
corresponding identifier or \code{ckan_group}. For users, pass a username
or \code{ckan_user}.}

\item{url}{Base URL to use. Default: https://demo.ckan.org/. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key. Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
the table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible. The result
can vary from a vector, list or data.frame. (required)}

\item{...}{Extra curl arguments. The function passes them to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
Create or remove follower relationships for a CKAN object. You must
authenticate as a user with an API key. The key represents the follower.
}
\examples{
\dontrun{
ckanr_setup(url = "https://demo.ckan.org/", key = "my-key")
dataset_follow("my-dataset")
dataset_am_following("my-dataset")
dataset_unfollow("my-dataset")
}
}
