% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/followers.R
\name{follow_users}
\alias{follow_users}
\alias{follow_user}
\alias{unfollow_user}
\alias{am_following_user}
\title{Follow or unfollow users}
\usage{
follow_user(
  id,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)

unfollow_user(
  id,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)

am_following_user(id, url = get_default_url(), key = get_default_key(), ...)
}
\arguments{
\item{id}{(character or \code{ckan_user}) The user to follow or inspect.}

\item{url}{Base URL to use. Default: https://demo.ckan.org/. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key. Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
the table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible. The result
can vary from a vector, list or data.frame. (required)}

\item{...}{Extra curl arguments. The function passes them to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
Manage user-to-user follower relationships.
}
\examples{
\dontrun{
ckanr_setup(url = "https://demo.ckan.org/", key = "my-key")
follow_user("demo-user")
am_following_user("demo-user")
unfollow_user("demo-user")
}
}
