% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocomplete.R
\name{format_autocomplete}
\alias{format_autocomplete}
\title{Autocomplete resource formats.}
\usage{
format_autocomplete(
  q,
  limit = 5,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\arguments{
\item{q}{(character) Partial string to search for. The function returns
entries whose name or title contains this string. (required)}

\item{limit}{(numeric) The maximum number of resource formats to return.
Default: 5 (optional)}

\item{url}{Base URL to use. Default: https://demo.ckan.org/. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key. Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
the table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible. The result
can vary from a vector, list or data.frame. (required)}

\item{...}{Extra curl arguments. The function passes them to \code{\link[crul]{verb-POST}} (optional)}
}
\value{
A character vector of matching format names. With \code{as = "table"},
a data.frame; with \code{as = "json"}, the raw JSON response.
}
\description{
Return a list of resource formats whose names contain a query string. See
\url{https://docs.ckan.org/en/2.11/api/#ckan.logic.action.get.format_autocomplete}
for the official API contract.
}
\examples{
\dontrun{
ckanr_setup(url = "https://demo.ckan.org/", key = getOption("ckan_demo_key"))

format_autocomplete(q = "csv")
}
}
\references{
https://docs.ckan.org/en/2.11/api/#ckan.logic.action.get.format_autocomplete
}
