% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_user.R
\name{get_site_user}
\alias{get_site_user}
\title{Show the CKAN site user.}
\usage{
get_site_user(
  defer_commit = FALSE,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\arguments{
\item{defer_commit}{(logical) By default (\code{FALSE}), \code{get_site_user}
commits and cleans up the current transaction. If \code{TRUE}, the caller is
responsible for committing the transaction after the call. Leaving open
connections can cause CLI commands to hang (optional).}

\item{url}{Base URL to use. Default: https://demo.ckan.org/. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key. Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
the table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible. The result
can vary from a vector, list or data.frame. (required)}

\item{...}{Extra curl arguments. The function passes them to \code{\link[crul]{verb-POST}} (optional)}
}
\value{
A \code{ckan_user} object. With \code{as = "table"}, a data.frame; with
\code{as = "json"}, the raw JSON response.
}
\description{
Return the CKAN site user. The site user is a special internal user that
the web interface and background jobs use for internal operations. You
must be a sysadmin to call this endpoint. See
\url{https://docs.ckan.org/en/2.11/api/#ckan.logic.action.get.get_site_user}
for the official API contract.
}
\examples{
\dontrun{
ckanr_setup(url = "https://demo.ckan.org/", key = getOption("ckan_demo_key"))

get_site_user()
}
}
\references{
https://docs.ckan.org/en/2.11/api/#ckan.logic.action.get.get_site_user
}
