% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/membership.R
\name{group_list_authz}
\alias{group_list_authz}
\title{List groups the current user can edit}
\usage{
group_list_authz(
  type = "group",
  groups = NULL,
  all_fields = FALSE,
  include_dataset_count = TRUE,
  include_extras = FALSE,
  include_tags = FALSE,
  include_groups = FALSE,
  include_users = FALSE,
  available_only = FALSE,
  am_member = FALSE,
  offset = NULL,
  limit = NULL,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\arguments{
\item{type}{(character) Group type, defaults to \code{"group"}.}

\item{groups}{(character vector) Optional list of group names to filter.}

\item{all_fields}{(logical) Return full group dictionaries.}

\item{include_dataset_count, include_extras, include_tags, include_groups, include_users}{(logical) Include additional metadata when \code{all_fields = TRUE}.}

\item{available_only}{(logical) Remove groups already associated to the context dataset.}

\item{am_member}{(logical) Only return memberships for the current user.}

\item{offset}{(numeric) Where to start getting activity items from
(optional, default: 0)}

\item{limit}{(numeric) The maximum number of activities to return
(optional, default: 31)}

\item{url}{Base URL to use. Default: https://demo.ckan.org/. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key. Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
the table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible. The result
can vary from a vector, list or data.frame. (required)}

\item{...}{Extra curl arguments. The function passes them to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
List groups the current user can edit
}
