% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/membership.R
\name{group_member_create}
\alias{group_member_create}
\title{Make a user a member of a group}
\usage{
group_member_create(
  id,
  username,
  role,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\arguments{
\item{id}{(character, \code{ckan_group}, or \code{ckan_organization}) Identifier for the container.}

\item{username}{(character or \code{ckan_user}) User to add.}

\item{role}{(character) Desired role (\code{"member"}, \code{"editor"}, \code{"admin"}).}

\item{url}{Base URL to use. Default: https://demo.ckan.org/. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key. Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
the table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible. The result
can vary from a vector, list or data.frame. (required)}

\item{...}{Extra curl arguments. The function passes them to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
Make a user a member of a group
}
