% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/membership.R
\name{member_delete}
\alias{member_delete}
\title{Remove a membership via member_* endpoints}
\usage{
member_delete(
  id,
  object,
  object_type,
  url = get_default_url(),
  key = get_default_key(),
  ...
)
}
\arguments{
\item{id}{(character, \code{ckan_group}, or \code{ckan_organization}) Identifier for the container.}

\item{object}{(character) The object to add (dataset id, user name, etc.).}

\item{object_type}{(character) Type of object you add (\code{"user"}, \code{"package"}, ...).}

\item{url}{Base URL to use. Default: https://demo.ckan.org/. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key. Default: your key set with
\code{\link{ckanr_setup}}}

\item{...}{Extra curl arguments. The function passes them to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
Remove a membership via member_* endpoints
}
