% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/membership.R
\name{member_roles_list}
\alias{member_roles_list}
\title{Report available member roles}
\usage{
member_roles_list(
  group_type = "organization",
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\arguments{
\item{group_type}{(character) Either \code{"group"} or \code{"organization"}.}

\item{url}{Base URL to use. Default: https://demo.ckan.org/. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key. Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
the table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible. The result
can vary from a vector, list or data.frame. (required)}

\item{...}{Extra curl arguments. The function passes them to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
Report available member roles
}
