% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization_purge.R
\name{organization_purge}
\alias{organization_purge}
\title{Purge an organization}
\usage{
organization_purge(
  id,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\arguments{
\item{id}{(character) name or id of the organization}

\item{url}{Base URL to use. Default: https://demo.ckan.org/. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key. Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
the table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible. The result
can vary from a vector, list or data.frame. (required)}

\item{...}{Extra curl arguments. The function passes them to \code{\link[crul]{verb-POST}} (optional)}
}
\value{
The function returns an empty list on success
}
\description{
IMPORTANT: You must be a sysadmin to purge an organization. Once you purge
an organization, the system removes it permanently.
}
\examples{
\dontrun{
ckanr_setup(url = "https://demo.ckan.org", key = getOption("ckan_demo_key"))

# create an organization
(res <- organization_create("foobar",
  title = "Foo bars",
  description = "love foo bars"
))

# delete the organization just created
res$id
organization_delete(id = res$id)

# purge the organization just deleted
res$id
organization_purge(id = res$id)
}
}
