% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization_update.R
\name{organization_update}
\alias{organization_update}
\title{Update an organization.}
\usage{
organization_update(
  x,
  id,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\arguments{
\item{x}{(list) A list with key-value pairs}

\item{id}{(character or \code{ckan_organization}) The name or id of the
organization to update, or a \code{ckan_organization} object.}

\item{url}{Base URL to use. Default: https://demo.ckan.org/. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key. Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
the table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible. The result
can vary from a vector, list or data.frame. (required)}

\item{...}{Extra curl arguments. The function passes them to \code{\link[crul]{verb-POST}} (optional)}
}
\value{
The updated organization as a \code{ckan_organization} object.
With \code{as = "table"}, a data.frame; with \code{as = "json"}, the raw JSON
response.
}
\description{
This function updates all organization metadata fields. You must be
authorized to edit the organization. Update methods may delete parameters
not explicitly provided: if you want to edit only specific attributes,
use \code{\link[=organization_patch]{organization_patch()}} instead. For the full list of accepted fields,
see \code{\link[=organization_create]{organization_create()}} and
\url{https://docs.ckan.org/en/2.11/api/#ckan.logic.action.update.organization_update}.
}
\examples{
\dontrun{
# Setup
ckanr_setup(url = "https://demo.ckan.org/", key = getOption("ckan_demo_key"))

# First, create an organization
org <- organization_create("water-bears2")
organization_show(org)

# Make some changes
x <- list(description = "An organization about water bears")

# Then update the organization
organization_update(x, id = org)
}
}
\references{
https://docs.ckan.org/en/2.11/api/#ckan.logic.action.update.organization_update
}
