% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_collaborators.R
\name{package_collaborator_delete}
\alias{package_collaborator_delete}
\title{Delete a dataset collaborator}
\usage{
package_collaborator_delete(
  id,
  user_id,
  url = get_default_url(),
  key = get_default_key(),
  ...
)
}
\arguments{
\item{id}{(character or \code{ckan_package}) Dataset identifier or object.}

\item{user_id}{(character or \code{ckan_user}) User identifier or object.}

\item{url}{Base URL to use. Default: https://demo.ckan.org/. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key. Default: your key set with
\code{\link{ckanr_setup}}}

\item{...}{Extra curl arguments. The function passes them to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
Delete a dataset collaborator
}
\examples{
\dontrun{
ckanr_setup(url = "https://demo.ckan.org/", key = "my-key")
package_collaborator_delete("my-dataset", "new-user")
}
}
