% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_collaborators.R
\name{package_collaborator_list}
\alias{package_collaborator_list}
\title{List collaborators for a dataset}
\usage{
package_collaborator_list(
  id,
  capacity = NULL,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\arguments{
\item{id}{(character or \code{ckan_package}) Dataset identifier or object.}

\item{capacity}{(character) Optional capacity filter (\code{"member"}, \code{"editor"}, \code{"admin"}).}

\item{url}{Base URL to use. Default: https://demo.ckan.org/. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key. Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
the table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible. The result
can vary from a vector, list or data.frame. (required)}

\item{...}{Extra curl arguments. The function passes them to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
List collaborators for a dataset
}
\examples{
\dontrun{
ckanr_setup(url = "https://demo.ckan.org/", key = "my-key")
package_collaborator_list("my-dataset")
}
}
