% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_relationships.R
\name{package_relationship_update}
\alias{package_relationship_update}
\title{Update an existing dataset relationship}
\usage{
package_relationship_update(
  subject,
  object,
  relationship_type,
  comment = NULL,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\arguments{
\item{subject}{(character or \code{ckan_package}) Dataset that acts as the subject in the relationship.}

\item{object}{(character or \code{ckan_package}) Dataset that acts as the object in the relationship.}

\item{relationship_type}{(character) Relationship type (\code{"depends_on"}, \code{"derives_from"}, etc.).}

\item{comment}{(character) Optional note for the relationship.}

\item{url}{Base URL to use. Default: https://demo.ckan.org/. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key. Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
the table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible. The result
can vary from a vector, list or data.frame. (required)}

\item{...}{Extra curl arguments. The function passes them to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
Update an existing dataset relationship
}
