% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_relationships.R
\name{package_relationships_list}
\alias{package_relationships_list}
\title{List relationships for a dataset}
\usage{
package_relationships_list(
  id,
  id2 = NULL,
  relationship_type = NULL,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\arguments{
\item{id}{(character or \code{ckan_package}) The dataset to inspect.}

\item{id2}{(character or \code{ckan_package}) Optional secondary dataset identifier to scope results.}

\item{relationship_type}{(character) Filter on relationship type (see \code{\link[=package_relationship_create]{package_relationship_create()}}).}

\item{url}{Base URL to use. Default: https://demo.ckan.org/. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key. Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
the table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible. The result
can vary from a vector, list or data.frame. (required)}

\item{...}{Extra curl arguments. The function passes them to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
List relationships for a dataset
}
\examples{
\dontrun{
ckanr_setup(url = "https://demo.ckan.org/", key = "my-key")
package_relationships_list("my-dataset")
}
}
