% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_dataset_extras.R
\name{package_revise}
\alias{package_revise}
\title{Revise a dataset using match/filter/update semantics}
\usage{
package_revise(
  match = NULL,
  filter = NULL,
  update = NULL,
  include = NULL,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\arguments{
\item{match}{(list) Key/value pairs that identify the dataset to revise. Unless you use flattened keys,
this parameter is required.}

\item{filter}{(character or list) Patterns that describe fields to remove before the update runs.}

\item{update}{(list) Values to set after filtering. The values support flattened keys.}

\item{include}{(character or list) Optional patterns that delimit which fields the response returns.}

\item{url}{Base URL to use. Default: https://demo.ckan.org/. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key. Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
the table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible. The result
can vary from a vector, list or data.frame. (required)}

\item{...}{Extra curl arguments. The function passes them to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
Revise a dataset using match/filter/update semantics
}
\examples{
\dontrun{
package_revise(
  match = list(name = "source-dataset"),
  update = list(notes = "New description")
)
}
}
