% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ping.R
\name{ping}
\alias{ping}
\title{Ping a CKAN server to test if it is up or down.}
\usage{
ping(url = get_default_url(), key = get_default_key(), as = "logical", ...)
}
\arguments{
\item{url}{Base URL to use. Default: https://demo.ckan.org/. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key. Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of "logical" (default) or "json". With
\code{as = "logical"} failures return \code{FALSE}. With \code{as = "json"} failures
signal an error instead of returning a non-JSON logical.}

\item{...}{Extra curl arguments. The function passes them to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
Ping a CKAN server to test if it is up or down.
}
\examples{
\dontrun{
ping()
ping(as = "json")
}
}
