% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_create.R
\name{resource_create}
\alias{resource_create}
\title{Create a resource}
\usage{
resource_create(
  package_id = NULL,
  rcurl = NULL,
  revision_id = NULL,
  description = NULL,
  format = NULL,
  hash = NULL,
  name = NULL,
  resource_type = NULL,
  mimetype = NULL,
  mimetype_inner = NULL,
  webstore_url = NULL,
  cache_url = NULL,
  size = NULL,
  created = NULL,
  last_modified = NULL,
  cache_last_updated = NULL,
  webstore_last_updated = NULL,
  upload = NULL,
  extras = NULL,
  http_method = "GET",
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\arguments{
\item{package_id}{(character) ID of the package. You add the resource
to this package. The value must be an alphanumeric string. Required.}

\item{rcurl}{(character) URL of the resource. Required.}

\item{revision_id}{(character) Revision ID. Optional.}

\item{description}{(character) Description of the resource. Optional. Required.}

\item{format}{(character) Format. Optional.}

\item{hash}{(character) Hash. Optional.}

\item{name}{(character) Name of the resource. Optional. Required.}

\item{resource_type}{(character) Resource type. Optional.}

\item{mimetype}{(character) MIME type. Optional.}

\item{mimetype_inner}{(character) Inner MIME type. Optional.}

\item{webstore_url}{(character) Webstore URL. Optional.}

\item{cache_url}{(character) Cache URL. Optional.}

\item{size}{(integer) Size. Optional.}

\item{created}{(character) ISO date string. Optional.}

\item{last_modified}{(character) ISO date string. Optional.}

\item{cache_last_updated}{(character) ISO date string. Optional.}

\item{webstore_last_updated}{(character) ISO date string. Optional.}

\item{upload}{(character) Path to a local file. Optional.}

\item{extras}{(list) Extra metadata fields of the resource. Optional.}

\item{http_method}{(character) HTTP method (verb) to use.
The value is one of "GET" or "POST". Default is "GET".}

\item{url}{Base URL to use. Default: https://demo.ckan.org/. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key. Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
the table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible. The result
can vary from a vector, list or data.frame. (required)}

\item{...}{Extra curl arguments. The function passes them to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
Create a resource
}
\examples{
\dontrun{
# Setup
ckanr_setup(
  url = "https://demo.ckan.org/",
  key = getOption("ckan_demo_key")
)

# create a package
(res <- package_create("foobarrrr", author = "Jane Doe"))

# then create a resource
file <- system.file("examples", "actinidiaceae.csv", package = "ckanr")
(xx <- resource_create(
  package_id = res$id,
  description = "my resource",
  name = "bears",
  upload = file,
  extras = list(species = "grizzly"),
  rcurl = "http://google.com"
))

package_create("foobbbbbarrrr") \%>\%
  resource_create(
    description = "my resource",
    name = "bearsareus",
    upload = file,
    extras = list(my_extra = "some value"),
    rcurl = "http://google.com"
  )
}
}
