% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_views.R
\name{resource_create_default_resource_views}
\alias{resource_create_default_resource_views}
\title{Create default views for a single resource}
\usage{
resource_create_default_resource_views(
  resource,
  package = NULL,
  create_datastore_views = FALSE,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\arguments{
\item{resource}{(character or \code{ckan_resource}) Resource identifier or object.}

\item{package}{(optional) Dataset identifier or \code{ckan_package} object.}

\item{create_datastore_views}{(logical) When \code{TRUE}, the function creates only
views that require DataStore-backed resources. The function uses this option
when DataPusher finishes ingesting data.}

\item{url}{Base URL to use. Default: https://demo.ckan.org/. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key. Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
the table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible. The result
can vary from a vector, list or data.frame. (required)}

\item{...}{Extra curl arguments. The function passes them to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
Create default views for a single resource
}
