% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_get.R
\name{resource_get}
\alias{resource_get}
\title{Get a resource.}
\usage{
resource_get(id, url = get_default_url(), key = get_default_key(),
  as = "list", ...)
}
\arguments{
\item{id}{(character) Resource identifier.}

\item{url}{Base url to use. Default: \url{http://data.techno-science.ca}. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{as}{(character) One of list (default), table, or json. Parsing with table option
uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)}, which attempts to parse
data to data.frame's when possible, so the result can vary from a vector, list or
data.frame. (required)}

\item{...}{Curl args passed on to \code{\link[httr]{POST}} (optional)}
}
\description{
Get a resource.
}
\examples{
\dontrun{
# Setup
ckanr_setup(url = "http://demo.ckan.org/", key = getOption("ckan_demo_key"))

# create a package
(res <- package_create("yellow7"))

# then create a resource
file <- system.file("examples", "actinidiaceae.csv", package = "ckanr")
(xx <- resource_create(package_id = res$id,
                       description = "my resource",
                       name = "bears",
                       upload = file,
                       rcurl = "http://google.com"
))

# downloads the resource
resource_get(xx$id)

# process downloaded file, e.g. read as table
x <- resource_get(xx$id)
table <- read.table(textConnection(x))

}
}
