% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_search.R
\name{resource_search}
\alias{resource_search}
\title{Search for resources.}
\usage{
resource_search(
  q,
  sort = NULL,
  offset = NULL,
  limit = NULL,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\arguments{
\item{q}{Query terms. The value is a string of the form \code{field:term}
or a vector or list of strings. Each string uses the same form.
\code{field} is a field or extra field on the Resource domain object.
If \code{field} is hash, the function matches the \code{term} as a prefix
of the Resource.hash field. If \code{field} is an extra field, the function
matches against the extra fields stored against the Resource.}

\item{sort}{Field to sort on. You can specify ascending (for example,
score desc) or descending (for example, score asc). You can sort by two
fields (for example, score desc, price asc). You can sort by a function
(for example, sum(x_f, y_f) desc). The function sorts by the sum of x_f
and y_f in descending order.}

\item{offset}{Record to start at. The default is the beginning.}

\item{limit}{Number of records to return.}

\item{url}{Base URL to use. Default: https://demo.ckan.org/. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key. Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
the table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible. The result
can vary from a vector, list or data.frame. (required)}

\item{...}{Extra curl arguments. The function passes them to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
Search for resources.
}
\examples{
\dontrun{
resource_search(q = "name:data")
resource_search(q = "name:data", as = "json")
resource_search(q = "name:data", as = "table")
resource_search(q = "name:data", limit = 2, as = "table")
resource_search(q = c("description:encoded", "name:No.2"), url = "demo.ckan.org")
}
}
