% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_views.R
\name{resource_view_clear}
\alias{resource_view_clear}
\title{Clear resource views}
\usage{
resource_view_clear(
  view_types = NULL,
  url = get_default_url(),
  key = get_default_key(),
  ...
)
}
\arguments{
\item{view_types}{(character vector) Optional subset of view types to delete.
When \code{NULL}, the function removes all views.}

\item{url}{Base URL to use. Default: https://demo.ckan.org/. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key. Default: your key set with
\code{\link{ckanr_setup}}}

\item{...}{Extra curl arguments. The function passes them to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
Clear resource views
}
