% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_views.R
\name{resource_view_delete}
\alias{resource_view_delete}
\title{Delete a resource view}
\usage{
resource_view_delete(id, url = get_default_url(), key = get_default_key(), ...)
}
\arguments{
\item{id}{(character or \code{ckan_resource_view}) View identifier.}

\item{url}{Base URL to use. Default: https://demo.ckan.org/. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key. Default: your key set with
\code{\link{ckanr_setup}}}

\item{...}{Extra curl arguments. The function passes them to \code{\link[crul]{verb-POST}} (optional)}
}
\value{
(bool) Result of the deletion. The value is TRUE when the function
deletes the resource view successfully, and FALSE when it does not.
}
\description{
Delete a resource view
}
