% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_views.R
\name{resource_view_update}
\alias{resource_view_update}
\title{Update an existing resource view}
\usage{
resource_view_update(
  id,
  resource = NULL,
  title = NULL,
  description = NULL,
  config = NULL,
  filter_fields = NULL,
  filter_values = NULL,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\arguments{
\item{id}{(character or \code{ckan_resource_view}) View identifier.}

\item{resource}{(character or \code{ckan_resource}) Parent resource identifier.}

\item{title}{(character) Title of the view.}

\item{description}{(character) Description. Optional.}

\item{config}{(list) Arbitrary configuration list. The function passes it
to the plugin.}

\item{filter_fields, filter_values}{Optional filter parameters for filterable views.}

\item{url}{Base URL to use. Default: https://demo.ckan.org/. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key. Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
the table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible. The result
can vary from a vector, list or data.frame. (required)}

\item{...}{Extra curl arguments. The function passes them to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
Update an existing resource view
}
