% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servers.R
\name{servers}
\alias{servers}
\title{CKAN server URLs and other info}
\usage{
servers()
}
\description{
CKAN server URLs and other info
}
\details{
CKAN instances from https://github.com/ckan/ckan-instances/
}
\examples{
\dontrun{
servers()
ckan_info(servers()[5])

# what version is each CKAN server running
out <- lapply(servers()[1:6], function(w) {
  ckan_version(w)$version_num
})
}
}
