% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_key.R
\name{set_api_key}
\alias{set_api_key}
\title{Set API KEY of CKAN}
\usage{
set_api_key(api_key)
}
\arguments{
\item{api_key}{(character). The api key of the user.}
}
\description{
Some CKAN API functions require authentication.
The \code{set_api_key} is a helper function to set the api key and
enable authentication.
}
\details{
The \code{ckanr} will check the value of global options \code{X-CKAN-API-Key}
before communicating with CKAN server. If the value is not \code{NULL}, the
\code{ckanr} will use the value as api key to authenticate with CKAN service.
}
\examples{
\dontrun{
set_api_key("xxx")
# list the private resources authorized to the above api key
organization_show("some_name", include_datasets = TRUE)[["packages"]]
}
}
\seealso{
Other ckanr.options: \code{\link{ckanr_options}};
  \code{\link{get_ckanr_url}}, \code{\link{set_ckanr_url}}
}

