% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ckanr_options.R
\name{set_test_env}
\alias{set_test_env}
\title{Set options for CKAN test environment}
\usage{
set_test_env(url, key, did, rid)
}
\arguments{
\item{url}{A valid CKAN URL for testing purposes}

\item{key}{A valid CKAN API key privileged to create datasets at URL}

\item{did}{A valid CKAN dataset ID, existing at url}

\item{rid}{A valid CKAN resource ID, attached to did}
}
\description{
Tests require a valid CKAN URL, a privileged API key for that URL,
 plus the IDs of an existing dataset and an existing resource, repectively.
}
\details{
The settings are written to both options and Sys.env. Testthat can only
 access Sys.env.
}

