% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ckanr_settings.R
\name{setup_ckanr}
\alias{setup_ckanr}
\title{Configure default CKAN settings}
\usage{
# An unprivileged read-only user could run either of:
setup_ckanr("http://data-demo.dpaw.wa.gov.au/")
setup_ckanr(url="http://data-demo.dpaw.wa.gov.au/")

# A privileged CKAN editor/admin user can run either of:
setup_ckanr("http://data-demo.dpaw.wa.gov.au/", "some-CKAN-API-key")
setup_ckanr(url="http://data-demo.dpaw.wa.gov.au/", key="some-CKAN-API-key")
set_ckanr_url("http://data-demo.dpaw.wa.gov.au/"); set_api_key("some-CKAN-api-key")

# ckanR developer/tester can run either of:
setup_ckanr("http://data-demo.dpaw.wa.gov.au/", "some-CKAN-API-key",
           "http://test-ckan.dpaw.wa.gov.au/","test-ckan-API-key",
           "test-ckan-dataset-id","test-ckan-resource-id")
setup_ckanr(url="http://data-demo.dpaw.wa.gov.au/", key="some-CKAN-API-key",
           test_url="http://test-ckan.dpaw.wa.gov.au/",test_key="test-ckan-API-key",
           test_did="test-ckan-dataset-id",test_rid="test-ckan-resource-id")

set_ckanr_url("http://data-demo.dpaw.wa.gov.au/"); set_api_key("some-CKAN-api-key")
set_test_env(url="http://test-ckan.dpaw.wa.gov.au/", key="test-ckan-API-key",
             did="test-ckan-dataset-id", rid="test-ckan-resource-id")

# This will set the CKAN URL to its default:
setup_ckanr()
}
\arguments{
\item{url}{A CKAN URL (optional), default: "http://data.techno-science.ca/"}

\item{key}{A CKAN API key (optional)}

\item{test_url}{(optional) A valid CKAN URL for testing purposes}

\item{test_key}{(optional) A valid CKAN API key privileged to create datasets
at \code{test_url}}

\item{test_did}{(optional) A valid CKAN dataset ID, existing at \code{test_url}}

\item{test_rid}{(optional) A valid CKAN resource ID, attached to \code{did}}
}
\description{
Configure default CKAN settings
}

