% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_list.R
\name{tag_list}
\alias{tag_list}
\title{List tags.}
\usage{
tag_list(
  query = NULL,
  vocabulary_id = NULL,
  all_fields = FALSE,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\arguments{
\item{query}{(character) Tag name query to search for. If you give a query,
the function returns only tags whose names contain this string.}

\item{vocabulary_id}{(character) ID or name of a vocabulary.
If you give a vocabulary, the function returns only tags that belong
to this vocabulary.}

\item{all_fields}{(logical) The function returns full tag dictionaries
instead of names. Default is \code{FALSE}.}

\item{url}{Base URL to use. Default: https://demo.ckan.org/. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key. Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
the table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible. The result
can vary from a vector, list or data.frame. (required)}

\item{...}{Extra curl arguments. The function passes them to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
List tags.
}
\examples{
\dontrun{
# list all tags
tag_list()

# search for a specific tag
tag_list(query = "aviation")

# all fields
tag_list(all_fields = TRUE)

# give back different data formats
tag_list("aviation", as = "json")
tag_list("aviation", as = "table")
}
}
