% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_search.R
\name{tag_search}
\alias{tag_search}
\title{Search tags.}
\usage{
tag_search(
  query = NULL,
  vocabulary_id = NULL,
  offset = 0,
  limit = 31,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\arguments{
\item{query}{(character) Tag name query to search for. If you give a query,
the function returns only tags whose names contain this string. The value
is one or more search strings.}

\item{vocabulary_id}{(character) ID or name of a vocabulary.
If you give a vocabulary, the function returns only tags that belong
to this vocabulary.}

\item{offset}{(numeric) Where to start getting activity items from
(optional, default: 0)}

\item{limit}{(numeric) The maximum number of activities to return
(optional, default: 31)}

\item{url}{Base URL to use. Default: https://demo.ckan.org/. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key. Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
the table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible. The result
can vary from a vector, list or data.frame. (required)}

\item{...}{Extra curl arguments. The function passes them to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
Search tags.
}
\examples{
\dontrun{
tag_search(query = "ta")
tag_search(query = c("ta", "al"))

# different formats back
tag_search(query = "ta", as = "json")
tag_search(query = "ta", as = "table")
}
}
