% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{tbl.src_ckan}
\alias{tbl.src_ckan}
\title{An implementation of \code{dplyr::tbl()} for \code{ckanr}}
\usage{
\method{tbl}{src_ckan}(src, name, ...)
}
\arguments{
\item{src}{A \code{DBI::dbConnect()} database connection.}

\item{name}{The table name which will be used as the SQL FROM clause.
This is vulnerable to SQL injections, so use only with known safe names,
e.g. CKAN resource IDs.}

\item{...}{Extra arguments will be passed on to \code{dbplyr::tbl_sql()}.}
}
\description{
This implementation converts the \code{DBI::dbConnect()} connection returned by
\code{ckanr::src_ckan()} into a structure that nests the connection under a key
\code{con}. This is expected by \code{dbplyr::tbl_sql()}.
}
