% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_show.R
\name{user_show}
\alias{user_show}
\title{Show a user.}
\usage{
user_show(
  id,
  user_obj = NULL,
  include_datasets = FALSE,
  include_num_followers = FALSE,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\arguments{
\item{id}{(character) User identifier.}

\item{user_obj}{(user dictionary) User dictionary of the user. Optional.}

\item{include_datasets}{(logical) Include a list of datasets that the user
creates. If the same user or a sysadmin requests the data, the function
includes datasets that are draft or private. Optional. Default is False.
The limit is 50.}

\item{include_num_followers}{(logical) Include the number of followers
of the user. Optional. Default is False.}

\item{url}{Base URL to use. Default: https://demo.ckan.org/. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key. Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
the table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible. The result
can vary from a vector, list or data.frame. (required)}

\item{...}{Extra curl arguments. The function passes them to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
Show a user.
}
\examples{
\dontrun{
# Setup
ckanr_setup(url = "https://demo.ckan.org/", key = getOption("ckan_demo_key"))

# show user
user_show("sckottie")

# include datasets
user_show("sckottie", include_datasets = TRUE)

# include datasets
user_show("sckottie", include_num_followers = TRUE)
}
}
