% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{activity_helpers}
\alias{activity_helpers}
\alias{group_activity_list}
\alias{organization_activity_list}
\alias{recently_changed_packages_activity_list}
\alias{dashboard_new_activities_count}
\alias{dashboard_mark_activities_old}
\alias{activity_show}
\alias{activity_data_show}
\alias{activity_diff}
\alias{activity_create}
\alias{send_email_notifications}
\title{Activity stream helpers}
\usage{
group_activity_list(
  id,
  offset = 0,
  limit = 31,
  include_hidden_activity = FALSE,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)

organization_activity_list(
  id,
  offset = 0,
  limit = 31,
  include_hidden_activity = FALSE,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)

recently_changed_packages_activity_list(
  offset = 0,
  limit = 31,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)

dashboard_new_activities_count(
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)

dashboard_mark_activities_old(
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)

activity_show(
  id,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)

activity_data_show(
  id,
  object_type,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)

activity_diff(
  id,
  object_type,
  diff_type = "unified",
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)

activity_create(
  user_id,
  object_id,
  activity_type,
  data = NULL,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)

send_email_notifications(
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\description{
These helpers wrap CKAN's \code{ckanext.activity} endpoints. They require the
\code{activity} core plugin to be enabled on the target CKAN instance.
}
