% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_key.R
\name{api_key}
\alias{api_key}
\title{Enable Authentication with API-KEY}
\usage{
api_key()
}
\description{
Some CKAN API functions require authentication.
The \code{set_api_key} is a helper function to set the API-KEY and
the \code{api_key} is a helper function to enable authentication.
}
\keyword{internal}

