% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as-ckan_organization.R
\name{as.ckan_organization}
\alias{as.ckan_organization}
\alias{is.ckan_organization}
\title{ckan_organization class helpers}
\usage{
as.ckan_organization(x, ...)

is.ckan_organization(x)
}
\arguments{
\item{x}{Variety of things, character, list, or ckan_organization class object}

\item{...}{Further args passed on to \code{\link{organization_show}} if character given}
}
\description{
ckan_organization class helpers
}
\examples{
\dontrun{
ckanr_setup(url = "http://demo.ckan.org/", key = getOption("ckan_demo_key"))

(orgs <- organization_list())
orgs[[3]]

# create item class from only an item ID
as.ckan_organization(orgs[[3]]$id)

# gives back itself
(x <- as.ckan_organization(orgs[[3]]$id))
as.ckan_organization(x)
}
}

