% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as-ckan_package.R
\name{as.ckan_package}
\alias{as.ckan_package}
\alias{is.ckan_package}
\title{ckan_package class helpers}
\usage{
as.ckan_package(x, ...)

is.ckan_package(x)
}
\arguments{
\item{x}{Variety of things, character, list, or ckan_package class object}

\item{...}{Further args passed on to \code{\link{package_show}} if character given}
}
\description{
ckan_package class helpers
}
\examples{
\dontrun{
ckanr_setup(url = "http://demo.ckan.org/", key = getOption("ckan_demo_key"))

(pkgs <- package_search())
pkgs$results
pkgs$results[[3]]

# create item class from only an item ID
as.ckan_package("0699f475-6978-473a-8448-42585074b6f1")

# gives back itself
(x <- as.ckan_package("0699f475-6978-473a-8448-42585074b6f1"))
as.ckan_package(x)
}
}

