% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-ckan_related.R
\name{as.ckan_related}
\alias{as.ckan_related}
\alias{is.ckan_related}
\title{ckan_related class helpers}
\usage{
as.ckan_related(x, ...)

is.ckan_related(x)
}
\arguments{
\item{x}{Variety of things, character, list, or ckan_related class object}

\item{...}{Further args passed on to \code{\link{related_show}} if character given}
}
\description{
ckan_related class helpers
}
\examples{
\dontrun{
ckanr_setup(url = "https://demo.ckan.org/", key = getOption("ckan_demo_key"))

(x <- package_create("foobbbbbarrrr") \%>\%
   related_create(title = "my resource",
                  type = "visualization"))

# create item class from only an item ID
as.ckan_related(x$id)

# gives back itself
(x <- as.ckan_related(x$id))
as.ckan_related(x)
}
}
