% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-ckan_resource_view.R
\name{as.ckan_resource_view}
\alias{as.ckan_resource_view}
\alias{is.ckan_resource_view}
\title{ckan_resource_view class helpers}
\usage{
as.ckan_resource_view(x, ...)

is.ckan_resource_view(x)
}
\arguments{
\item{x}{Variety of things: character, list, or \code{ckan_resource_view} object}

\item{...}{Further arguments passed to \code{\link[=resource_view_show]{resource_view_show()}} when \code{x} is an
identifier.}
}
\description{
ckan_resource_view class helpers
}
\examples{
\dontrun{
ckanr_setup(url = "https://demo.ckan.org/",
  key = Sys.getenv("CKAN_DEMO_KEY"))

res <- package_show("sample-dataset")
views <- resource_view_list(res$resources[[1]]$id)

# Coerce from ID
as.ckan_resource_view(views[[1]]$id)

# Pass through existing objects
as.ckan_resource_view(views[[1]])
}
}
