% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as-ckan_user.R
\name{as.ckan_user}
\alias{as.ckan_user}
\alias{is.ckan_user}
\title{ckan_user class helpers}
\usage{
as.ckan_user(x, ...)

is.ckan_user(x)
}
\arguments{
\item{x}{Variety of things, character, list, or ckan_user class object}

\item{...}{Further args passed on to \code{\link{user_show}} if character given}
}
\description{
ckan_user class helpers
}
\examples{
\dontrun{
ckanr_setup(url = "http://demo.ckan.org/", key = getOption("ckan_demo_key"))

(usrs <- user_list())
usrs$results
usrs$results[[3]]

# create item class from only an item ID
as.ckan_user(usrs$results[[3]]$id)

# gives back itself
(x <- as.ckan_user(usrs$results[[3]]$id))
as.ckan_user(x)
}
}

