% Generated by roxygen2 (4.0.2): do not edit by hand
\name{changes}
\alias{changes}
\title{Get an activity stream of recently changed datasets on a site.}
\usage{
changes(offset = 0, limit = 31, url = "http://data.techno-science.ca",
  as = "list", ...)
}
\arguments{
\item{offset}{(numeric) Where to start getting activity items from (optional, default: 0)}

\item{limit}{(numeric) The maximum number of activities to return (optional, default: 31)}

\item{url}{Base url to use. Default: \url{http://data.techno-science.ca}}

\item{as}{(character) One of list (default), table, or json. Parsing with table option
uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)}, which attempts to parse
data to data.frame's when possible, so the result can vary.}

\item{...}{Curl args passed on to \code{\link[httr]{POST}}}
}
\description{
Get an activity stream of recently changed datasets on a site.
}
\examples{
\donttest{
changes()
changes(as='json')
changes(as='table')
}
}

