% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changes.R
\name{changes}
\alias{changes}
\title{Get an activity stream of recently changed datasets on a site.}
\usage{
changes(offset = 0, limit = 31, url = get_default_url(),
  as = "list", ...)
}
\arguments{
\item{offset}{(numeric) Where to start getting activity items from (optional, default: 0)}

\item{limit}{(numeric) The maximum number of activities to return (optional, default: 31)}

\item{url}{Base url to use. Default: \url{http://data.techno-science.ca}. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{as}{(character) One of list (default), table, or json. Parsing with table option
uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)}, which attempts to parse
data to data.frame's when possible, so the result can vary from a vector, list or
data.frame. (required)}

\item{...}{Curl args passed on to \code{\link[httr]{POST}} (optional)}
}
\description{
Get an activity stream of recently changed datasets on a site.
}
\examples{
\dontrun{
changes()
changes(as = 'json')
changes(as = 'table')
}
}
