% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-action.R
\name{ckan_action}
\alias{ckan_action}
\title{Generic action function}
\usage{
ckan_action(
  action,
  query = NULL,
  body = NULL,
  headers = list(),
  verb = "POST",
  url = get_default_url(),
  key = get_default_key(),
  ...
)
}
\arguments{
\item{action}{A valid CKAN API action name (e.g., \code{"package_list"}, \code{"package_show"}). See the CKAN API documentation for a full list of actions: \url{https://docs.ckan.org/en/latest/api/index.html}}

\item{query}{a named list of URL query parameters}

\item{body}{data to be used in the body of a request, see
https://docs.ropensci.org/crul/reference/verb-POST.html for options}

\item{headers}{a named list of request headers}

\item{verb}{HTTP request verb, e.g., GET, POST}

\item{url}{Base url to use. Default: https://data.ontario.ca/ See
also \code{\link[=ckanr_setup]{ckanr_setup()}} and \code{\link[=get_default_url]{get_default_url()}}}

\item{key}{A privileged CKAN API key, Default: your key set with
\code{\link{ckanr_setup}}}

\item{...}{Curl args passed on to the appropriate crul HTTP verb function (e.g., \link[crul:verb-GET]{crul::verb-GET}, \link[crul:verb-POST]{crul::verb-POST}) depending on the value of \code{verb} (optional)}
}
\value{
a text string. because we can't know ahead of time for sure
what kind of data will be returned, we return text and the user
can parse as needed.
}
\description{
You have to set the CKAN action to use, as well as the HTTP verb,
in addition to any query parameters, request body data, and
request headers.
}
\examples{
\dontrun{
ckanr_setup(url = "https://demo.ckan.org/",
 key = getOption("ckan_demo_key"))

ckan_action("package_list")
ckan_action("package_list", verb = "GET")
ckan_action("package_list", url = 'https://data.nhm.ac.uk')
}
}
