% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ckan_info.R
\name{ckan_info}
\alias{ckan_info}
\title{Get information on a CKAN server}
\usage{
ckan_info(url = get_default_url(), ...)
}
\arguments{
\item{url}{Base url to use. Default: \url{http://data.techno-science.ca}. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}. (required)}

\item{...}{Curl args passed on to \code{\link[httr]{GET}} (optional)}
}
\description{
Get information on a CKAN server
}
\examples{
\dontrun{
ckan_info()
ckan_info(servers()[5])
}
}

