% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckan_info.R
\name{ckan_info}
\alias{ckan_info}
\alias{ckan_version}
\title{Get information on a CKAN server}
\usage{
ckan_info(url = get_default_url(), ...)

ckan_version(url = get_default_url(), ...)
}
\arguments{
\item{url}{Base url to use. Default: \url{https://data.ontario.ca}. See
also \code{\link[=ckanr_setup]{ckanr_setup()}} and \code{\link[=get_default_url]{get_default_url()}}. (required)}

\item{...}{Curl args passed on to \link[crul:verb-GET]{crul::verb-GET} (optional)}
}
\value{
for \code{ckan_info} a list with many slots with various info.
for \code{ckan_version}, list of length two, with actual version as character,
and another with version converted to numeric (any dots or letters removed)
}
\description{
Get information on a CKAN server
}
\examples{
\dontrun{
ckan_info()
ckan_info(servers()[5])

ckan_version(servers()[5])
}
}
