% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ckanr-package.R
\docType{package}
\name{ckanr-package}
\alias{ckanr}
\alias{ckanr-package}
\title{R client for the CKAN API}
\description{
ckanr is a full client for the CKAN API, wrapping all
APIs, including for reading and writing data. Please get in touch
(\url{https://github.com/ropensci/ckanr/issues}) if you have problems, or
have use cases that we don't cover yet.
}
\section{CKAN API}{


Document for the CKAN API is at \url{http://docs.ckan.org/en/latest/api/index.html}.
We'll always be following the lastest version of the API.
}

\section{ckanr package API}{


The functions can be grouped into those for setup, packages,
resources, tags, organizations, and groups.

\itemize{
 \item Setup - The main one is \code{\link{ckanr_setup}} - and many related
 functions, e.g., \code{\link{get_default_key}}
 \item Packages - Create a package with \code{\link{package_create}}, and see
 other functions starting with \code{package_*}
 \item Resources - Create a package with \code{\link{resource_create}}, and see
 other functions starting with \code{resource_*}
 \item Tags - List tags with \code{\link{tag_list}}, and see
 other functions starting with \code{tag_*}
 \item Organizations - List organizations with \code{\link{organization_list}},
 and show a specific organization with \code{\link{organization_show}}
}
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}

Imanuel Costigan \email{i.costigan@me.com}

Wush Wu

Florian Mayer \email{florian.wendelin.mayer@gmail.com}
}
\keyword{package}

