% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ckanr-package.R
\docType{package}
\name{ckanr-package}
\alias{ckanr}
\alias{ckanr-package}
\title{R client for the CKAN API}
\description{
R client for the CKAN API
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

