% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ckanr_options.R
\name{ckanr_options}
\alias{ckanr_options}
\title{Get ckanr options}
\usage{
ckanr_options()
}
\value{
Prints your base url, API key (if used), and optional test server
 settings (URL, API key, a dataset ID and a resource ID)
}
\description{
Get ckanr options
}
\examples{
ckanr_options()
}
\seealso{
\code{\link{set_ckanr_url}}, \code{\link{get_ckanr_url}},
\code{\link{set_api_key}}

Other ckanr.options: \code{\link{get_ckanr_url}},
  \code{\link{set_ckanr_url}}; \code{\link{set_api_key}}
}

