% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ckanr_settings.R
\name{ckanr_settings}
\alias{ckanr_settings}
\alias{get_default_key}
\alias{get_default_url}
\alias{get_test_did}
\alias{get_test_key}
\alias{get_test_rid}
\alias{get_test_url}
\title{Get or set ckanr CKAN settings}
\usage{
ckanr_settings()

get_default_url()

get_default_key()

get_test_url()

get_test_key()

get_test_did()

get_test_rid()
}
\value{
\code{ckanr_settings} prints your base url, API key (if used), and
optional test server settings (URL, API key, a dataset ID and a resource ID).
\code{set_test_env} sets your test settings, while \code{get_test_url},
\code{get_test_key}, \code{get_test_did}, and \code{get_test_rid} get
each of those respective settings.
}
\description{
Get or set ckanr CKAN settings
}
\details{
\code{set_test_env}: Set options for CKAN test environment. Tests require a
valid CKAN URL, a privileged API key for that URL, plus the IDs of an existing
dataset and an existing resource, repectively. The settings are written as
environment variables.
}
\examples{
ckanr_settings()
}
\seealso{
\code{\link{set_ckanr_url}}, \code{\link{get_ckanr_url}},
\code{\link{set_api_key}}
}

