% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_count.R
\name{dashboard_count}
\alias{dashboard_count}
\title{Number of new activities of an authorized user}
\usage{
dashboard_count(url = get_default_url(), key = get_default_key(),
  as = "list", ...)
}
\arguments{
\item{url}{Base url to use. Default: \url{http://data.techno-science.ca}. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key, Default: your key set with \code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with table option
uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)}, which attempts to parse
data to data.frame's when possible, so the result can vary from a vector, list or
data.frame. (required)}

\item{...}{Curl args passed on to \code{\link[httr]{POST}} (optional)}
}
\description{
Number of new activities of an authorized user
}
\details{
Important: Activities from the user herself are not counted by this
function even though they appear in the dashboard (users don't want to be
notified about things they did themselves).
}
\examples{
\dontrun{
# Setup
ckanr_setup(url = "https://demo.ckan.org/", key = getOption("ckan_demo_key"))

# count
dashboard_count()
}
}
