% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ds_create.R
\name{ds_create}
\alias{ds_create}
\title{Datastore - create a datastore}
\usage{
ds_create(url = "http://demo.ckan.org", as = "list", ...)
}
\arguments{
\item{url}{Base url to use. Default: \url{http://data.techno-science.ca}}

\item{as}{(character) One of list (default), table, or json. Parsing with table option
uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)}, which attempts to parse
data to data.frame's when possible, so the result can vary.}

\item{...}{Curl args passed on to \code{\link[httr]{POST}}}
}
\description{
Datastore - create a datastore
}
\examples{
\donttest{
ds_create(id='6b30b22e-2bd2-4e3e-b64f-c67d4fb4e3be', url = 'http://demo.ckan.org/')
}
}

