% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ds_search_sql.R
\name{ds_search_sql}
\alias{ds_search_sql}
\title{Datastore - search or get a dataset from CKRAN datastore}
\usage{
ds_search_sql(sql, url = "http://demo.ckan.org", as = "list", ...)
}
\arguments{
\item{sql}{(character) A single SQL select statement. (required)}

\item{url}{Base url to use. Default: \url{http://data.techno-science.ca}. See
also \code{\link{get_ckanr_url}} and \code{\link{set_ckanr_url}}}

\item{as}{(character) One of list (default), table, or json. Parsing with table option
uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)}, which attempts to parse
data to data.frame's when possible, so the result can vary. (required)}

\item{...}{Curl args passed on to \code{\link[httr]{POST}} (optional)}
}
\description{
Datastore - search or get a dataset from CKRAN datastore
}
\examples{
\donttest{
url <- 'http://demo.ckan.org/'
sql = 'SELECT * from "f4129802-22aa-4437-b9f9-8a8f3b7b2a53" LIMIT 2'
ds_search_sql(sql, url=url, as="table")
sql2 = 'SELECT "Species","Genus","Family" from "f4129802-22aa-4437-b9f9-8a8f3b7b2a53" LIMIT 2'
ds_search_sql(sql2, url=url, as="table")
}
}

