% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/followers.R
\name{followee_lists}
\alias{followee_lists}
\alias{followee_list}
\alias{user_followee_list}
\alias{dataset_followee_list}
\alias{group_followee_list}
\alias{organization_followee_list}
\title{Followee lists for CKAN users}
\usage{
followee_list(
  id,
  q = NULL,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)

user_followee_list(
  id,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)

dataset_followee_list(
  id,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)

group_followee_list(
  id,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)

organization_followee_list(
  id,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\arguments{
\item{id}{(character or \code{ckan_user}) User identifier.}

\item{q}{(character) Optional prefix filter for \code{followee_list()}.}

\item{url}{Base URL to use. Default: https://demo.ckan.org/. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key. Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
the table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible. The result
can vary from a vector, list or data.frame. (required)}

\item{...}{Extra curl arguments. The function passes them to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
The function returns the objects that a user follows.

See
\url{https://docs.ckan.org/en/2.11/api/#ckan.logic.action.get.organization_followee_list}
for the official API contract.
}
\examples{
\dontrun{
ckanr_setup(url = "https://demo.ckan.org/", key = "my-key")
followee_list("demo-user", q = "ckan")
dataset_followee_list("demo-user")
organization_followee_list("demo-user")
}
}
