% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/followers.R
\name{follower_counts}
\alias{follower_counts}
\alias{dataset_follower_count}
\alias{group_follower_count}
\alias{user_follower_count}
\title{Follower counts for CKAN objects}
\usage{
dataset_follower_count(
  id,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)

group_follower_count(
  id,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)

user_follower_count(
  id,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\arguments{
\item{id}{(character or \verb{ckan_*}) Identifier for the object to inspect.
For datasets pass an id/slug or \code{ckan_package}, for groups pass the
corresponding identifier or \code{ckan_group}, and for users pass a username
or \code{ckan_user}.}

\item{url}{Base url to use. Default: https://demo.ckan.org/ See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key, Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible, so the result
can vary from a vector, list or data.frame. (required)}

\item{...}{Curl args passed on to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
Report follower totals for datasets, groups, or users.
}
\examples{
\dontrun{
ckanr_setup(url = "https://demo.ckan.org/", key = "my-key")
dataset_follower_count("my-dataset")
group_follower_count("my-group")
user_follower_count("demo-user")
}
}
