% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_show.R
\name{group_show}
\alias{group_show}
\title{Show a group}
\usage{
group_show(
  id,
  include_datasets = TRUE,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\arguments{
\item{id}{(character) Package identifier.}

\item{include_datasets}{(logical) Include a list of the group's datasets.
Default: \code{TRUE}}

\item{url}{Base URL to use. Default: https://demo.ckan.org/. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key. Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
the table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible. The result
can vary from a vector, list or data.frame. (required)}

\item{...}{Extra curl arguments. The function passes them to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
Show a group
}
\details{
By default the function drops the help and success slots. It returns
only the result slot. If you want raw json, request \code{as = 'json'}.
Then you parse the result yourself to get the help slot.
}
\examples{
\dontrun{
res <- group_list()

# via a group name/id
group_show(res[[1]]$name)

# or via an object of class ckan_group
group_show(res[[1]])

# return different data formats
group_show(res[[1]]$name, as = "json")
group_show(res[[1]]$name, as = "table")
}
}
