% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_ops.R
\name{job_queue}
\alias{job_queue}
\alias{job_list}
\alias{job_show}
\alias{job_clear}
\alias{job_cancel}
\title{Background job helpers}
\usage{
job_list(
  queues = NULL,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)

job_show(
  id,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)

job_clear(
  queues = NULL,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)

job_cancel(
  id,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\arguments{
\item{queues}{(character) Queue names to target.}

\item{url}{Base URL to use. Default: https://demo.ckan.org/. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key. Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
the table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible. The result
can vary from a vector, list or data.frame. (required)}

\item{...}{Extra curl arguments. The function passes them to \code{\link[crul]{verb-POST}} (optional)}

\item{id}{(character) Job identifier.}
}
\description{
Manage the RQ/Redis job queue of CKAN. You must have sysadmin access.
}
\examples{
\dontrun{
ckanr_setup(url = "https://demo.ckan.org/", key = "my-ckan-key")
job_list()
}
}
