% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/license_list.R
\name{license_list}
\alias{license_list}
\title{Return the list of licenses available for datasets on the site.}
\usage{
license_list(id, url = get_default_url(), as = "list", ...)
}
\arguments{
\item{id}{(character) Package identifier.}

\item{url}{Base url to use. Default: \url{http://data.techno-science.ca}. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{as}{(character) One of list (default), table, or json. Parsing with table option
uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)}, which attempts to parse
data to data.frame's when possible, so the result can vary from a vector, list or
data.frame. (required)}

\item{...}{Curl args passed on to \code{\link[httr]{POST}} (optional)}
}
\description{
Return the list of licenses available for datasets on the site.
}
\examples{
\dontrun{
license_list()
license_list(as = "table")
license_list(as = "json")
}
}

