% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/membership.R
\name{member_create}
\alias{member_create}
\title{Create or update a membership via member_* endpoints}
\usage{
member_create(
  id,
  object,
  object_type,
  capacity,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\arguments{
\item{id}{(character, \code{ckan_group}, or \code{ckan_organization}) Identifier for the container.}

\item{object}{(character) The object to add (dataset id, user name, etc.).}

\item{object_type}{(character) Type of object being added (\code{"user"}, \code{"package"}, ...).}

\item{capacity}{(character) Capacity for the membership.}

\item{url}{Base url to use. Default: https://data.ontario.ca/ See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key, Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible, so the result
can vary from a vector, list or data.frame. (required)}

\item{...}{Curl args passed on to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
Create or update a membership via member_* endpoints
}
