% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/membership.R
\name{member_list}
\alias{member_list}
\title{List members for a group or organization}
\usage{
member_list(
  id,
  object_type = NULL,
  capacity = NULL,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\arguments{
\item{id}{(character, \code{ckan_group}, or \code{ckan_organization}) Identifier for the container.}

\item{object_type}{(character) Optional object type filter (\code{"user"}, \code{"package"}, etc.).}

\item{capacity}{(character) Optional capacity filter (\code{"member"}, \code{"editor"}, \code{"admin"}).}

\item{url}{Base url to use. Default: https://data.ontario.ca/ See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key, Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible, so the result
can vary from a vector, list or data.frame. (required)}

\item{...}{Curl args passed on to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
List members for a group or organization
}
\examples{
# ckanr_setup(url = "https://demo.ckan.org/", key = "my-key")
# member_list("my-group", object_type = "user")
}
