% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization_delete.R
\name{organization_delete}
\alias{organization_delete}
\title{Delete an organization}
\usage{
organization_delete(id, key = get_default_key(),
  url = get_default_url(), as = "list", ...)
}
\arguments{
\item{id}{(character) name or id of the organization}

\item{key}{A privileged CKAN API key, Default: your key set with \code{\link{ckanr_setup}}}

\item{url}{Base url to use. Default: \url{http://data.techno-science.ca}. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{as}{(character) One of list (default), table, or json. Parsing with table option
uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)}, which attempts to parse
data to data.frame's when possible, so the result can vary from a vector, list or
data.frame. (required)}

\item{...}{Curl args passed on to \code{\link[httr]{POST}} (optional)}
}
\description{
Delete an organization
}
\examples{
\dontrun{
ckanr_setup(url = "https://demo.ckan.org/")

# delete an organization
organization_delete(id = "someorgid")
}
}
