% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/organization_list.R
\name{organization_list}
\alias{organization_list}
\title{List organization}
\usage{
organization_list(order_by = c("name", "package"), decreasing = FALSE,
  organizations = NULL, all_fields = TRUE, url = get_ckanr_url(),
  as = "list", ...)
}
\arguments{
\item{order_by}{(character, only the first element is used).
The field to sort the list by, must be \code{name} or \code{packages}.}

\item{decreasing}{(logical). Is the sort-order is decreasing or not.}

\item{organizations}{NOT IMPLEMENTED YET}

\item{all_fields}{(logical). Return the name or all fields of the object.}

\item{url}{Base url to use. Default: \url{http://data.techno-science.ca}. See
also \code{\link{get_ckanr_url}} and \code{\link{set_ckanr_url}}}

\item{as}{(character) One of list (default), table, or json. Parsing with table option
uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)}, which attempts to parse
data to data.frame's when possible, so the result can vary. (required)}

\item{...}{Curl args passed on to \code{\link[httr]{POST}} (optional)}
}
\description{
List organization
}
\examples{
\donttest{
organization_list()
organization_list(as='json')
organization_list(as='table')
}
}

