% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/membership.R
\name{organization_list_for_user}
\alias{organization_list_for_user}
\title{List organizations a user can act on}
\usage{
organization_list_for_user(
  id = NULL,
  permission = "manage_group",
  include_dataset_count = FALSE,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\arguments{
\item{id}{(character or \code{ckan_user}) Optional user identifier; defaults to the
authenticated user.}

\item{permission}{(character) Permission to check (defaults to \code{"manage_group"}).}

\item{include_dataset_count}{(logical) Include dataset counts in the response.}

\item{url}{Base url to use. Default: https://demo.ckan.org/ See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key, Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible, so the result
can vary from a vector, list or data.frame. (required)}

\item{...}{Curl args passed on to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
List organizations a user can act on
}
