% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/membership.R
\name{organization_member_delete}
\alias{organization_member_delete}
\title{Remove a user from an organization}
\usage{
organization_member_delete(
  id,
  username,
  url = get_default_url(),
  key = get_default_key(),
  ...
)
}
\arguments{
\item{id}{(character, \code{ckan_group}, or \code{ckan_organization}) Identifier for the container.}

\item{username}{(character or \code{ckan_user}) User to add.}

\item{url}{Base url to use. Default: https://data.ontario.ca/ See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key, Default: your key set with
\code{\link{ckanr_setup}}}

\item{...}{Curl args passed on to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
Remove a user from an organization
}
