% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package_activity_list.R
\name{package_activity_list}
\alias{package_activity_list}
\title{Return a list of the package's activity}
\usage{
package_activity_list(id, offset = 0, limit = 31, url = get_default_url(),
  as = "list", ...)
}
\arguments{
\item{id}{(character) Package identifier.}

\item{offset}{(numeric) Where to start getting activity items from (optional, default: 0)}

\item{limit}{(numeric) The maximum number of activities to return (optional, default: 31)}

\item{url}{Base url to use. Default: \url{http://data.techno-science.ca}. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{as}{(character) One of list (default), table, or json. Parsing with table option
uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)}, which attempts to parse
data to data.frame's when possible, so the result can vary. (required)}

\item{...}{Curl args passed on to \code{\link[httr]{POST}} (optional)}
}
\description{
Return a list of the package's activity
}
\examples{
\dontrun{
package_activity_list("34d60b13-1fd5-430e-b0ec-c8bc7f4841cf")
package_activity_list("34d60b13-1fd5-430e-b0ec-c8bc7f4841cf", as = "table")
package_activity_list("34d60b13-1fd5-430e-b0ec-c8bc7f4841cf", as = "json")
}
}

