% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_collaborators.R
\name{package_collaborator_create}
\alias{package_collaborator_create}
\title{Create or update a dataset collaborator}
\usage{
package_collaborator_create(
  id,
  user_id,
  capacity,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\arguments{
\item{id}{(character or \code{ckan_package}) Dataset identifier or object.}

\item{user_id}{(character or \code{ckan_user}) User identifier or object.}

\item{capacity}{(character) Collaborator role (\code{"member"}, \code{"editor"}, \code{"admin"}).}

\item{url}{Base url to use. Default: https://data.ontario.ca/ See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key, Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible, so the result
can vary from a vector, list or data.frame. (required)}

\item{...}{Curl args passed on to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
Create or update a dataset collaborator
}
