% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_views.R
\name{package_create_default_resource_views}
\alias{package_create_default_resource_views}
\title{Create default views for all resources in a dataset}
\usage{
package_create_default_resource_views(
  package,
  create_datastore_views = FALSE,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\arguments{
\item{package}{(character or \code{ckan_package}) Dataset identifier.}

\item{create_datastore_views}{(logical) When \code{TRUE}, only create views that
require DataStore-backed resources (used when DataPusher finishes ingesting data).}

\item{url}{Base url to use. Default: https://data.ontario.ca/ See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key, Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible, so the result
can vary from a vector, list or data.frame. (required)}

\item{...}{Curl args passed on to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
Create default views for all resources in a dataset
}
