% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package_list_current.R
\name{package_list_current}
\alias{package_list_current}
\title{List current packages with resources.}
\usage{
package_list_current(offset = 0, limit = 31, url = get_default_url(),
  as = "list", ...)
}
\arguments{
\item{offset}{(numeric) Where to start getting activity items from (optional, default: 0)}

\item{limit}{(numeric) The maximum number of activities to return (optional, default: 31)}

\item{url}{Base url to use. Default: \url{http://data.techno-science.ca}. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{as}{(character) One of list (default), table, or json. Parsing with table option
uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)}, which attempts to parse
data to data.frame's when possible, so the result can vary from a vector, list or
data.frame. (required)}

\item{...}{Curl args passed on to \code{\link[httr]{POST}} (optional)}
}
\description{
List current packages with resources.
}
\examples{
\dontrun{
package_list_current()
package_list_current(as = 'json')
package_list_current(as = 'table')
}
}

