% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_dataset_extras.R
\name{package_owner_org_update}
\alias{package_owner_org_update}
\title{Move a dataset to another organization}
\usage{
package_owner_org_update(
  id,
  organization_id,
  url = get_default_url(),
  key = get_default_key(),
  ...
)
}
\arguments{
\item{id}{(character or \code{ckan_package}) Dataset identifier.}

\item{organization_id}{(character or \code{ckan_organization}) Owning organization identifier.}

\item{url}{Base URL to use. Default: https://demo.ckan.org/. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key. Default: your key set with
\code{\link{ckanr_setup}}}

\item{...}{Extra curl arguments. The function passes them to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
Move a dataset to another organization
}
\examples{
\dontrun{
package_owner_org_update("dataset-id", organization_id = "target-org")
}
}
