% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_relationships.R
\name{package_relationship_delete}
\alias{package_relationship_delete}
\title{Delete a dataset relationship}
\usage{
package_relationship_delete(
  subject,
  object,
  relationship_type,
  url = get_default_url(),
  key = get_default_key(),
  ...
)
}
\arguments{
\item{subject}{(character or \code{ckan_package}) Dataset acting as the subject in the relationship.}

\item{object}{(character or \code{ckan_package}) Dataset acting as the object in the relationship.}

\item{relationship_type}{(character) Relationship type (\code{"depends_on"}, \code{"derives_from"}, etc.).}

\item{url}{Base url to use. Default: https://data.ontario.ca/ See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key, Default: your key set with
\code{\link{ckanr_setup}}}

\item{...}{Curl args passed on to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
Delete a dataset relationship
}
