% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_relationships.R
\name{package_relationships}
\alias{package_relationships}
\title{Package relationship helpers}
\description{
Utilities for managing dataset relationships via CKAN's \verb{package_relationship_*} endpoints.
}
