% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/package_relationships_list.R
\name{package_relationships_list}
\alias{package_relationships_list}
\title{Return a list of the site’s user accounts.}
\usage{
package_relationships_list(id, id2, url = "http://data.techno-science.ca",
  as = "list", ...)
}
\arguments{
\item{url}{Base url to use. Default: \url{http://data.techno-science.ca}}

\item{as}{(character) One of list (default), table, or json. Parsing with table option
uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)}, which attempts to parse
data to data.frame's when possible, so the result can vary.}

\item{...}{Curl args passed on to \code{\link[httr]{POST}}}

\item{q}{(character) Restrict the users returned to those whose names contain a string}

\item{order_by}{(character) Which field to sort the list by (optional, default: 'name')}
}
\description{
Return a list of the site’s user accounts.
}
\examples{
\donttest{
id="e179e910-27fb-44f4-a627-99822af49ffa"
id2="a4233ee2-a675-4c6a-803c-177e6ab220c0"
package_relationships_list(id, id2)
package_relationships_list(as="table")
package_relationships_list(as="json")
}
}

