% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_dataset_extras.R
\name{package_resource_reorder}
\alias{package_resource_reorder}
\title{Reorder resources for a dataset}
\usage{
package_resource_reorder(
  id,
  order,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\arguments{
\item{id}{(character or \code{ckan_package}) Dataset identifier.}

\item{order}{(character vector) Resource IDs in the desired order.}

\item{url}{Base url to use. Default: https://data.ontario.ca/ See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key, Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible, so the result
can vary from a vector, list or data.frame. (required)}

\item{...}{Curl args passed on to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
Reorder resources for a dataset
}
\examples{
\dontrun{
pkg <- package_show("my-dataset")
package_resource_reorder(pkg,
  order = rev(vapply(pkg$resources, `[[`, "id", FUN.VALUE = character(1)))
)
}
}
