% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package_revision_list.R
\name{package_revision_list}
\alias{package_revision_list}
\title{Return a dataset (package's) revisions as a list of dictionaries.}
\usage{
package_revision_list(id, url = get_default_url(), as = "list", ...)
}
\arguments{
\item{id}{(character) Package identifier.}

\item{url}{Base url to use. Default: \url{http://data.techno-science.ca}. See
also \code{\link{setup_ckanr}} and \code{\link{get_default_url}}.}

\item{as}{(character) One of list (default), table, or json. Parsing with table option
uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)}, which attempts to parse
data to data.frame's when possible, so the result can vary. (required)}

\item{...}{Curl args passed on to \code{\link[httr]{POST}} (optional)}
}
\description{
Return a dataset (package's) revisions as a list of dictionaries.
}
\examples{
\dontrun{
package_revision_list('34d60b13-1fd5-430e-b0ec-c8bc7f4841cf')
package_revision_list('34d60b13-1fd5-430e-b0ec-c8bc7f4841cf', as = "table")
package_revision_list('34d60b13-1fd5-430e-b0ec-c8bc7f4841cf', as = "json")
}
}

