% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package_show.R
\name{package_show}
\alias{package_show}
\title{Show a package.}
\usage{
package_show(id, use_default_schema = FALSE, url = get_ckanr_url(),
  as = "list", ...)
}
\arguments{
\item{id}{(character) Package identifier.}

\item{use_default_schema}{(logical) Use default package schema instead of a custom
schema defined with an IDatasetForm plugin. Default: FALSE}

\item{url}{Base url to use. Default: \url{http://data.techno-science.ca}. See
also \code{\link{get_ckanr_url}} and \code{\link{set_ckanr_url}}}

\item{as}{(character) One of list (default), table, or json. Parsing with table option
uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)}, which attempts to parse
data to data.frame's when possible, so the result can vary. (required)}

\item{...}{Curl args passed on to \code{\link[httr]{POST}} (optional)}
}
\description{
Show a package.
}
\details{
By default the help and success slots are dropped, and only the result
slot is returned. You can request raw json with \code{as='json'} then parse yourself
to get the help slot.
}
\examples{
\donttest{
package_show('34d60b13-1fd5-430e-b0ec-c8bc7f4841cf')
package_show('34d60b13-1fd5-430e-b0ec-c8bc7f4841cf', as='json')
package_show('34d60b13-1fd5-430e-b0ec-c8bc7f4841cf', as='table')
package_show('34d60b13-1fd5-430e-b0ec-c8bc7f4841cf', TRUE)
}
}

