% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package_update.R
\name{package_update}
\alias{package_update}
\title{Update a package}
\usage{
package_update(id, url = get_default_url(), as = "list", ...)
}
\arguments{
\item{id}{(character) Package identifier.}

\item{url}{Base url to use. Default: \url{http://data.techno-science.ca}. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{as}{(character) One of list (default), table, or json. Parsing with table option
uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)}, which attempts to parse
data to data.frame's when possible, so the result can vary. (required)}

\item{...}{Curl args passed on to \code{\link[httr]{POST}} (optional)}

\item{use_default_schema}{(logical) Use default package schema instead of a
custom schema defined with an IDatasetForm plugin. Default: FALSE}
}
\description{
Update a package
}
\details{
By default the help and success slots are dropped, and only the
  result slot is returned. You can request raw json with \code{as = 'json'}
  then parse yourself to get the help slot.
}
\examples{
\dontrun{
# Setup
ckanr_setup(url = "http://demo.ckan.org/", key = getOption("ckan_demo_key"))

# First, show a package
res <- package_show('34d60b13-1fd5-430e-b0ec-c8bc7f4841cf')

# Make some changes
res$author <- "Jane Doe"

# Then update the packge
package_update(res)
}
}

