% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ping.R
\name{ping}
\alias{ping}
\title{Ping a CKAN server to test that it's up or down.}
\usage{
ping(url = get_ckanr_url(), as = "logical", ...)
}
\arguments{
\item{url}{Base url to use. Default: \url{http://data.techno-science.ca}}

\item{as}{(character) One of logical (default) or json. logical returns a logical, json
returns json.}

\item{...}{Curl args passed on to \code{\link[httr]{POST}}}
}
\description{
Ping a CKAN server to test that it's up or down.
}
\examples{
\donttest{
ping()
ping(as="json")
}
}

