% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ping.R
\name{ping}
\alias{ping}
\title{Ping a CKAN server to test that it's up or down.}
\usage{
ping(url = get_default_url(), as = "logical", ...)
}
\arguments{
\item{url}{Base url to use. Default: \url{http://data.techno-science.ca}. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{as}{(character) One of list (default), table, or json. Parsing with table option
uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)}, which attempts to parse
data to data.frame's when possible, so the result can vary from a vector, list or
data.frame. (required)}

\item{...}{Curl args passed on to \code{\link[httr]{POST}} (optional)}
}
\description{
Ping a CKAN server to test that it's up or down.
}
\examples{
\dontrun{
ping()
ping(as = "json")
}
}

