% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/related_create.R
\name{related_create}
\alias{related_create}
\title{Create a related item}
\usage{
related_create(
  id,
  title,
  type,
  description = NULL,
  related_id = NULL,
  related_url = NULL,
  image_url = NULL,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\arguments{
\item{id}{(character) ID of the package. You add the related item
to this package. The value must be an alphanumeric string. Required.}

\item{title}{(character) Title of the related item. Required.}

\item{type}{(character) Type of the related item. The value is one of API,
application, idea, news article, paper, post or visualization. Required.}

\item{description}{(character) Description of the related item. Optional.}

\item{related_id}{(character) ID to assign to the related item.
If the value is blank, the function assigns an ID. Optional.}

\item{related_url}{(character) URL for the related item.
Optional.}

\item{image_url}{(character) URL of an image for the related item. Optional.}

\item{url}{Base URL to use. Default: https://demo.ckan.org/. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key. Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
the table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible. The result
can vary from a vector, list or data.frame. (required)}

\item{...}{Extra curl arguments. The function passes them to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
Create a related item
}
\examples{
\dontrun{
# Setup
ckanr_setup(url = "https://demo.ckan.org/", key = getOption("ckan_demo_key"))

# create a package
(res <- package_create("hello-mars"))

# create a related item
related_create(res, title = "asdfdaf", type = "idea")

# pipe operations together
package_create("foobbbbbarrrr") \%>\%
  related_create(
    title = "my resource",
    type = "visualization"
  )
}
}
