% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/resource_search.R
\name{resource_search}
\alias{resource_search}
\title{Search for resources.}
\usage{
resource_search(q, sort = NULL, offset = NULL, limit = NULL,
  url = get_ckanr_url(), as = "list", ...)
}
\arguments{
\item{q}{Query terms, defaults to '*:*', or everything. It is a string of the form
\code{field:term} or a list of strings, each of the same form.  Within each string, \code{field}
is a field or extra field on the Resource domain object. If \code{field} is hash, then an
attempt is made to match the \code{term} as a *prefix* of the Resource.hash field. If
\code{field} is an extra field, then an attempt is made to match against the extra fields
stored against the Resource.}

\item{sort}{Field to sort on. You can specify ascending (e.g., score desc) or
descending (e.g., score asc), sort by two fields (e.g., score desc, price asc),
or sort by a function (e.g., sum(x_f, y_f) desc, which sorts by the sum of
x_f and y_f in a descending order).}

\item{offset}{Record to start at, default to beginning.}

\item{limit}{Number of records to return.}

\item{url}{Base url to use. Default: \url{http://data.techno-science.ca}. See
also \code{\link{get_ckanr_url}} and \code{\link{set_ckanr_url}}}

\item{as}{(character) One of list (default), table, or json. Parsing with table option
uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)}, which attempts to parse
data to data.frame's when possible, so the result can vary. (required)}

\item{...}{Curl args passed on to \code{\link[httr]{POST}} (optional)}
}
\description{
Search for resources.
}
\examples{
\donttest{
resource_search(q = 'name:data')
resource_search(q = 'name:data', as='json')
resource_search(q = 'name:data', as='table')
resource_search(q = 'name:data', limit = 2, as='table')
}
}

