% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/resource_show.R
\name{resource_show}
\alias{resource_show}
\title{Show a resource.}
\usage{
resource_show(id, url = get_ckanr_url(), as = "list", ...)
}
\arguments{
\item{id}{A resource ID}

\item{url}{Base url to use. Default: \url{http://data.techno-science.ca}. See
also \code{\link{get_ckanr_url}} and \code{\link{set_ckanr_url}}}

\item{as}{(character) One of list (default), table, or json. Parsing with table option
uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)}, which attempts to parse
data to data.frame's when possible, so the result can vary. (required)}

\item{...}{Curl args passed on to \code{\link[httr]{POST}} (optional)}
}
\description{
Show a resource.
}
\examples{
\donttest{
resource_show(id = "e179e910-27fb-44f4-a627-99822af49ffa", as="table")
}
}

