% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/resource_update.R
\name{resource_update}
\alias{resource_update}
\title{Update a resource's file attachment}
\usage{
resource_update(id, path, url = get_ckanr_url(),
  key = getOption("X-CKAN-API-Key"), as = "list", ...)
}
\arguments{
\item{id}{(character) Resource ID to update (required)}

\item{path}{(character) Local path of the file to upload (required)}

\item{url}{Base url to use. Default: \url{http://data.techno-science.ca}. See
also \code{\link{get_ckanr_url}} and \code{\link{set_ckanr_url}}}

\item{key}{A CKAN API key with write permissions to the resource's dataset
(default: \code{api_key()})}

\item{as}{(character) One of list (default), table, or json. Parsing with table option
uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)}, which attempts to parse
data to data.frame's when possible, so the result can vary. (required)}

\item{...}{Curl args passed on to \code{\link[httr]{POST}} (optional)}
}
\value{
The HTTP response from CKAN, formatted as list (default), table, or JSON.
}
\description{
This function will only update a resource's file attachment and the metadata
key "last_updated". Other metadata, such as name or description, are not
updated.
}
\details{
The new file must exist on a local path. R objects have to be written to a
file, e.g. using \code{tempfile()} - see example.

For convenience, CKAN base url and API key default to the global options,
which are set by \code{set_ckanr_url("http://my-ckan.org/")} and
\code{set_api_key("my-ckan-api-key")}, respectively.
}
\examples{
\dontrun{
# Using an existing file and explicit CKAN URL and API key
# Note: enter valid values for id, url, and key
resource_update(id="an-existing-resource-id",
                path=system.file("examples", "actinidiaceae.csv", package = "ckanr"),
                url="http://my-ckan-instance.org/", key="my-ckan-api-key")

# Using an R object written to a tempfile, and implicit CKAN URL and API key
data <- installed.packages()
path <- tempfile(fileext=".csv")
write.csv(data, path)
set_ckanr_url("http://demo.ckan.org/")
set_api_key("my-demo-ckan-org-api-key")
resource_update(id="an-existing-resource-id", path=path)
}
}
\references{
\url{http://docs.ckan.org/en/latest/api/index.html#ckan.logic.action.create.resource_create}
}

