% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_views.R
\name{resource_view_reorder}
\alias{resource_view_reorder}
\title{Reorder resource views}
\usage{
resource_view_reorder(
  id,
  order,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\arguments{
\item{id}{(character or \code{ckan_resource}) Resource identifier.}

\item{order}{(character vector) View IDs in the desired order.}

\item{url}{Base url to use. Default: https://demo.ckan.org/ See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key, Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible, so the result
can vary from a vector, list or data.frame. (required)}

\item{...}{Curl args passed on to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
Reorder resource views
}
