% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/revision_list.R
\name{revision_list}
\alias{revision_list}
\title{Return a list of the IDs of the site's revisions.}
\usage{
revision_list(url = "http://data.techno-science.ca", as = "list", ...)
}
\arguments{
\item{url}{Base url to use. Default: \url{http://data.techno-science.ca}}

\item{as}{(character) One of list (default), table, or json. Parsing with table option
uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)}, which attempts to parse
data to data.frame's when possible, so the result can vary.}

\item{...}{Curl args passed on to \code{\link[httr]{POST}}}
}
\description{
Return a list of the IDs of the site's revisions.
}
\examples{
\donttest{
revision_list()
revision_list(as="table")
revision_list(as="json")
}
}

