% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/servers.R
\name{servers}
\alias{servers}
\title{CKAN server URLS and other info}
\usage{
servers()
}
\description{
CKAN server URLS and other info
}
\examples{
servers()
ckan_info("http://data.gov.au/")
ckan_info(servers()[5])

# what version is each CKAN server running
out <- lapply(servers()[4:6], ckan_info)
vapply(out, "[[", "", "ckan_version")
}

