% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servers.R
\name{servers}
\alias{servers}
\title{CKAN server URLS and other info}
\usage{
servers()
}
\description{
CKAN server URLS and other info
}
\details{
Comes from the links at https://ckan.org/about/instances/

There were a number of other URLs for CKAN instances in the CKAN URL
above, but some sites are now gone completely, or if they do exist,
I can't figure out how to get access to the CKAN API on their instance.
}
\examples{
\dontrun{
servers()
ckan_info(servers()[5])

# what version is each CKAN server running
out <- lapply(servers()[1:6], ckan_info)
vapply(out, "[[", "", "ckan_version")
}
}
