% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/base_url.R
\name{set_ckanr_url}
\alias{get_ckanr_url}
\alias{set_ckanr_url}
\title{Set/get the default CKAN URL}
\usage{
set_ckanr_url(url = "http://data.techno-science.ca")

get_ckanr_url()
}
\arguments{
\item{url}{Base url to use. Default: \url{http://data.techno-science.ca}.}
}
\value{
sets the \code{ckanr.default.url} option with \code{\link{options}}.
}
\description{
The default CKAN URL is set to the \href{http://data.techno-science.ca}{CSTMC open data portal}
when the \code{ckanr} package is loaded. This base URL is then passed to
\code{ckanr} functions using \code{get_ckan_url()}.
}
\examples{
\dontrun{
get_ckanr_url()
set_ckanr_url("https://data.gov.au")
get_ckanr_url()
}
}

