% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabulary.R
\name{tag_autocomplete}
\alias{tag_autocomplete}
\title{Autocomplete tag names.}
\usage{
tag_autocomplete(
  q,
  vocabulary_id = NULL,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\arguments{
\item{q}{(character) Partial string to search for. The function returns
entries whose name or title contains this string. (required)}

\item{vocabulary_id}{(character) The function restricts matches to
this vocabulary.}

\item{url}{Base URL to use. Default: https://demo.ckan.org/. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key. Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
the table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible. The result
can vary from a vector, list or data.frame. (required)}

\item{...}{Extra curl arguments. The function passes them to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
Return a list of tag names that contain a query string. See
\url{https://docs.ckan.org/en/latest/api/} for the official API contract.
}
\examples{
\dontrun{
tag_autocomplete(q = "ckan")
}
}
