% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabulary.R
\name{tag_autocomplete}
\alias{tag_autocomplete}
\title{Autocomplete tag names.}
\usage{
tag_autocomplete(
  q,
  vocabulary_id = NULL,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\arguments{
\item{q}{(character) Partial tag name to search for. Required.}

\item{vocabulary_id}{(character) Restrict matches to a specific vocabulary.}

\item{url}{Base url to use. Default: https://demo.ckan.org/ See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key, Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible, so the result
can vary from a vector, list or data.frame. (required)}

\item{...}{Curl args passed on to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
Autocomplete tag names.
}
\examples{
\dontrun{
tag_autocomplete(q = "ckan")
}
}
