% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tag_create.R
\name{tag_create}
\alias{tag_create}
\title{Create a tag}
\usage{
tag_create(name, vocabulary_id, key = get_default_key(),
  url = get_default_url(), as = "list", ...)
}
\arguments{
\item{name}{(character) The name for the new tag, a string between 2 and 100
characters long containing only alphanumeric characters and -, _ and ., e.g. 'Jazz'}

\item{vocabulary_id}{(character) The id of the vocabulary that the new tag should be
added to, e.g. the id of vocabulary 'Genre'}

\item{key}{A privileged CKAN API key, Default: your key set with \code{\link{ckanr_setup}}}

\item{url}{Base url to use. Default: \url{http://data.techno-science.ca}. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{as}{(character) One of list (default), table, or json. Parsing with table option
uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)}, which attempts to parse
data to data.frame's when possible, so the result can vary from a vector, list or
data.frame. (required)}

\item{...}{Curl args passed on to \code{\link[httr]{POST}} (optional)}
}
\description{
Create a tag
}

