% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tag_list.R
\name{tag_list}
\alias{tag_list}
\title{List tags.}
\usage{
tag_list(query = NULL, vocabulary_id = NULL, all_fields = FALSE,
  url = "http://data.techno-science.ca", as = "list", ...)
}
\arguments{
\item{query}{(character) A tag name query to search for, if given only tags whose names contain
this string will be returned}

\item{vocabulary_id}{(character) The id or name of a vocabulary, if give only tags that belong
to this vocabulary will be returned}

\item{all_fields}{(logical) Return full tag dictionaries instead of just names. Default: FALSE}

\item{url}{Base url to use. Default: \url{http://data.techno-science.ca} (required)}

\item{as}{(character) One of list (default), table, or json. Parsing with table option
uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)}, which attempts to parse
data to data.frame's when possible, so the result can vary. (required)}

\item{...}{Curl args passed on to \code{\link[httr]{POST}} (optional)}
}
\description{
List tags.
}
\examples{
\donttest{
tag_list('aviation')
tag_list('aviation', as='json')
tag_list('aviation', as='table')
}
}

