% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tag_show.R
\name{tag_show}
\alias{tag_show}
\title{Show a tag.}
\usage{
tag_show(id, url = "http://data.techno-science.ca", as = "list", ...)
}
\arguments{
\item{id}{(character) The name or id of the tag}

\item{url}{Base url to use. Default: \url{http://data.techno-science.ca}}

\item{as}{(character) One of list (default), table, or json. Parsing with table option
uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)}, which attempts to parse
data to data.frame's when possible, so the result can vary.}

\item{...}{Curl args passed on to \code{\link[httr]{POST}}}
}
\description{
Show a tag.
}
\examples{
\donttest{
tag_show('Aviation')
tag_show('Aviation', as='json')
tag_show('Aviation', as='table')
}
}

