% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_create.R
\name{user_create}
\alias{user_create}
\title{Create a user.}
\usage{
user_create(
  name,
  email,
  password,
  id = NULL,
  fullname = NULL,
  about = NULL,
  openid = NULL,
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\arguments{
\item{name}{(character) Name of the new user. The name is a string between
2 and 100 characters in length. The name contains only lowercase alphanumeric
characters, - and _. Required.}

\item{email}{(character) Email address for the new user. Required.}

\item{password}{(character) Password of the new user. The password is a string
of at least 4 characters. Required.}

\item{id}{(character) ID of the new user. Optional.}

\item{fullname}{(character) Full name of the user. Optional.}

\item{about}{(character) Description of the new user. Optional.}

\item{openid}{(character) OpenID of the new user. Optional.}

\item{url}{Base URL to use. Default: https://demo.ckan.org/. See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key. Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
the table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible. The result
can vary from a vector, list or data.frame. (required)}

\item{...}{Extra curl arguments. The function passes them to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
Create a user.
}
\examples{
\dontrun{
# Setup
ckanr_setup(
  url = "https://data-demo.dpaw.wa.gov.au",
  key = "824e7c50-9577-4bfa-bf32-246ebed1a8a2"
)

# create a user
user_create(
  name = "stacy", email = "stacy@aaaaa.com",
  password = "helloworld"
)
}
}
\references{
http://docs.ckan.org/en/latest/api/index.html#ckan.logic.action.create.user_create
}
