% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/membership.R
\name{user_invite}
\alias{user_invite}
\title{Invite a user to a group}
\usage{
user_invite(
  email,
  group_id,
  role = "member",
  url = get_default_url(),
  key = get_default_key(),
  as = "list",
  ...
)
}
\arguments{
\item{email}{(character) Email address for the invitee.}

\item{group_id}{(character or \code{ckan_group}) Group identifier.}

\item{role}{(character) Desired role (\code{"member"}, \code{"editor"}, \code{"admin"}).}

\item{url}{Base url to use. Default: https://data.ontario.ca/ See
also \code{\link{ckanr_setup}} and \code{\link{get_default_url}}.}

\item{key}{A privileged CKAN API key, Default: your key set with
\code{\link{ckanr_setup}}}

\item{as}{(character) One of list (default), table, or json. Parsing with
table option uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)},
which attempts to parse data to data.frame's when possible, so the result
can vary from a vector, list or data.frame. (required)}

\item{...}{Curl args passed on to \code{\link[crul]{verb-POST}} (optional)}
}
\description{
Invite a user to a group
}
