% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/user_list.R
\name{user_list}
\alias{user_list}
\title{Return a list of the site's user accounts.}
\usage{
user_list(q = NULL, order_by = NULL,
  url = "http://data.techno-science.ca", as = "list", ...)
}
\arguments{
\item{q}{(character) Restrict the users returned to those whose names contain a string}

\item{order_by}{(character) Which field to sort the list by (optional, default: 'name')}

\item{url}{Base url to use. Default: \url{http://data.techno-science.ca} (required)}

\item{as}{(character) One of list (default), table, or json. Parsing with table option
uses \code{jsonlite::fromJSON(..., simplifyDataFrame = TRUE)}, which attempts to parse
data to data.frame's when possible, so the result can vary. (required)}

\item{...}{Curl args passed on to \code{\link[httr]{POST}} (optional)}
}
\description{
Return a list of the site's user accounts.
}
\examples{
\donttest{
user_list()
user_list(as="table")
user_list(as="json")
}
}

